Proarrow.Adjunction

Documentation

class (Profunctor p, Profunctor q) => Adjunction (p :: j +-> k) (q :: k +-> j) where

Adjunctions between two profunctors.

Methods

unit :: forall (a :: j). Ob a => (q :.: p) a a

counit :: (p :.: q) :~> ((~>) :: CAT k)

unit' :: forall {i} {j} (p :: i +-> j) (q :: j +-> i) (a :: i) (b :: i). Adjunction p q => (a ~> b) -> (q :.: p) a b

leftAdjunct :: forall {j} {k} (l :: j +-> k) r (a :: j) (b :: k). (Adjunction l r, Representable l, Representable r, Ob a) => ((l % a) ~> b) -> r a b

rightAdjunct :: forall {j} {k} (l :: j +-> k) r (a :: j) (b :: k). (Adjunction l r, Representable l, Representable r, Ob b) => r a b -> (l % a) ~> b

unitFromRepUnit :: forall {i} {j} (l :: i +-> j) (r :: j +-> i) (a :: i). (Representable l, Representable r, Ob a) => (a ~> (r % (l % a))) -> (r :.: l) a a

counitFromRepCounit :: forall {j} {k1} (l :: j +-> k1) (r :: k1 +-> j). (Representable l, Representable r) => (forall (c :: k1). Ob c => (l % (r % c)) ~> c) -> (l :.: r) :~> ((~>) :: CAT k1)

Orphan instances

(MonoidalProfunctor r, Adjunction l r, Representable l, Representable r, Monoidal j, Monoidal k) => MonoidalProfunctor (RepCostar l :: k -> j -> Type)

Instance details

Methods

par0 :: RepCostar l (Unit :: k) (Unit :: j)

par :: forall (x1 :: k) (x2 :: j) (y1 :: k) (y2 :: j). RepCostar l x1 x2 -> RepCostar l y1 y2 -> RepCostar l (x1 ** y1) (x2 ** y2)

Adjunction p q => Promonad (q :.: p :: k -> k -> Type)

Instance details

Methods

id :: forall (a :: k). Ob a => (q :.: p) a a

(.) :: forall (b :: k) (c :: k) (a :: k). (q :.: p) b c -> (q :.: p) a b -> (q :.: p) a c

Adjunction p q => Procomonad (p :.: q :: i -> i -> Type)

Instance details

Methods

extract :: (p :.: q) :~> ((~>) :: CAT i)

duplicate :: (p :.: q) :~> ((p :.: q) :.: (p :.: q))