Proarrow.Category.Bicategory

Bicategories

class BicategoryRequirements kk => Bicategory (kk :: CAT k) where

Bicategories. This is a strictified definition.

0-cells are kinds i, j, k... satisfying the Ob0 kk constraint.

1-cells are types of kind kk j k for any 0-cells j and k, satisfying the Ob1 kk constraint.

2-cells are values of type ps ~> qs, where ps and qs are of kind Path kk j k, representing lists of 1-cells.

Associated Types

type Ob0 kk (j :: k) :: Constraint

type Ob0 kk j = Any j

type Ob1 kk (p :: kk a b) :: Constraint

type Ob1 kk p = Any p

Methods

o :: ((ps :: Path kk i j) ~> qs) -> (rs ~> ss) -> (ps +++ rs) ~> (qs +++ ss)

Horizontal composition of 2-cells.

(\\\) :: ((Ob0 kk i, Ob0 kk j, Ob ps, Ob qs) => r) -> ((ps :: Path kk i j) ~> qs) -> r

Observe constraints from a 2-cell value.

type BICAT kk = forall {j} {k}. CAT (Path kk j k)

class (CategoryOf (Path kk j k), forall (ps :: Path kk j k). Ob ps => IsPath ps) => OkCategory kk j k

class Ob @(Path kk j j) Nil => PathNilIsOb kk j

id1 :: (Bicategory kk, Ob0 kk k) => (Nil :: Path kk k k) ~> (Nil :: Path kk k k)

The identity 1-cell, the unit of horizontal composition

obj1 :: forall {kk} {j} {k} (a :: kk j k). (Bicategory kk, Ob0 kk j, Ob0 kk k, Ob (a ::: Nil)) => Obj (a ::: Nil)

obj1 @a is short for obj @(a ::: Nil)

withAssoc :: forall {kk} {h} {i} {j} {k} (a :: Path kk h i) (b :: Path kk i j) (c :: Path kk j k) r. (Bicategory kk, Ob0 kk h, Ob0 kk i, Ob a, Ob b, Ob c) => (((a +++ b) +++ c) ~ (a +++ (b +++ c)) => r) -> r

withUnital :: forall {kk} {j} {k} (a :: Path kk j k) r. (Bicategory kk, Ob0 kk j, Ob0 kk k, Ob a) => ((a +++ Nil) ~ a => r) -> r

appendObj :: forall {kk} {i} {j} {k} (a :: Path kk i j) (b :: Path kk j k) r. (Bicategory kk, Ob0 kk i, Ob0 kk j, Ob0 kk k, Ob a, Ob b) => (Ob (a +++ b) => r) -> r

Paths

type data Path kk j k where

A type-level kind-threaded list. Used to strictify the bicategory and double category definitions. kk is a kind constructor, which creates a kind given two other kinds. (Each kind representing a 0-cell.)

Constructors

Nil :: Path kk k k

(:::) :: kk i j -> Path kk j k -> Path kk i k infixr 5

type MKKIND = CAT Kind

The type of 2-parameter kind constructors.

data SPath ps where

Constructors

SNil :: Ob0 kk k => SPath (Nil :: Path kk k k)

SCons :: forall {kk} {j} {k} (p :: kk j k) ps. (Ob0 kk j, Ob0 kk k, Ob1 kk p) => SPath ps -> SPath (p ::: ps)

class IsPath (ps :: Path kk j k) where

Methods

singPath :: SPath ps

withIsPath :: SPath ps -> (IsPath ps => r) -> r

type family (ps :: Path kk a b) +++ (qs :: Path kk b c) :: Path kk a c infixl 5

append :: SPath ps -> SPath qs -> SPath (ps +++ qs)

isPathAppend :: forall {kk} {j} {k} (ps :: Path kk j k) qs r. (IsPath ps, IsPath qs) => (IsPath (ps +++ qs) => r) -> r