Proarrow.Category.Bicategory.Relative

Documentation

class (Bicategory kk, Ob0 kk a, Ob0 kk e, Ob j, Ob t) => Monad (j :: kk e a) (t :: kk a e) where

A j-relative monad t. Note that j is the opposite of the usual convention. See Relative how to use this with a conjoint and a companion to get the regular definition.

Methods

unit :: (I :: kk a a) ~> O j t

mult :: O (O t j) t ~> t

class (Bicategory kk, Ob0 kk a, Ob0 kk d, Ob0 kk e, Ob j, Ob t) => Algebra (j :: kk e a) (t :: kk a e) (car :: kk d e) where

Methods

act :: O (O t j) car ~> car

class (Bicategory kk, Ob0 kk a, Ob0 kk b, Ob0 kk e, Ob j, Ob t) => Opalgebra (j :: kk e a) (t :: kk a e) (car :: kk a b) where

Methods

opact :: O (O car j) t ~> car

class (Bicategory kk, Ob0 kk a, Ob0 kk c, Ob0 kk e, Ob j, Ob l, Ob r) => Adjunction (j :: kk e a) (l :: kk a c) (r :: kk c e) where

Methods

eta :: (I :: kk a a) ~> O (O j r) l

epsilon :: O (O l j) r ~> (I :: kk c c)

class (Bicategory kk, Ob0 kk a, Ob0 kk e, Ob j, Ob t) => Comonad (j :: kk e a) (t :: kk a e) where

Methods

counit :: O j t ~> (I :: kk a a)

comult :: t ~> O (O t j) t

class (Bicategory kk, Ob0 kk a, Ob0 kk d, Ob0 kk e, Ob j, Ob t) => Coalgebra (j :: kk e a) (t :: kk a e) (car :: kk d e) where

Methods

coact :: car ~> O (O t j) car

class (Bicategory kk, Ob0 kk a, Ob0 kk b, Ob0 kk e, Ob j, Ob t) => Coopalgebra (j :: kk e a) (t :: kk a e) (car :: kk a b) where

Methods

coopact :: car ~> O (O car j) t

class (Bicategory kk, Ob0 kk a, Ob0 kk c, Ob0 kk e) => Coadjunction (j :: kk e a) (l :: kk a c) (r :: kk c e) where

Methods

coeta :: O (O j r) l ~> (I :: kk a a)

coepsilon :: (I :: kk c c) ~> O (O l j) r