Proarrow.Category.Bicategory.Strictified

Documentation

data Path (kk :: CAT k) (j :: k) (k1 :: k) where

The type of 2-parameter kind constructors.

A type-level kind-threaded list. Used to strictify the bicategory and double category definitions. kk is a kind constructor, which creates a kind given two other kinds. (Each kind representing a 0-cell.)

Constructors

Nil :: forall {k} (kk :: CAT k) (j :: k). Path kk j j

(:::) :: forall {k} (kk :: CAT k) (j :: k) (j1 :: k) (k1 :: k). kk j j1 -> Path kk j1 k1 -> Path kk j k1 infixr 5

type family (ps :: Path kk a b) +++ (qs :: Path kk b c) :: Path kk a c infixl 5

data SPath (ps :: Path kk j k1) where

Constructors

SNil :: forall {k} (kk :: CAT k) (j :: k). Ob0 kk j => SPath ('Nil :: Path kk j j)

SCons :: forall {k} {kk :: CAT k} {j :: k} {j1 :: k} {k1 :: k} (p :: kk j j1) (ps1 :: Path kk j1 k1). Ob0 kk j => Obj p -> SPath ps1 -> SPath (p '::: ps1)

class ((as +++ bs) +++ cs) ~ (as +++ (bs +++ cs)) => Assoc (as :: Path kk a a1) (bs :: Path kk a1 b) (cs :: Path kk b c)

class (Bicategory kk, Ob0 kk j, Ob0 kk k1, (ps +++ ('Nil :: Path kk k1 k1)) ~ ps, forall (i :: k) (h :: k) (qs :: Path kk k1 h) (rs :: Path kk h i). Assoc ps qs rs) => IsPath (ps :: Path kk j k1) where

Methods

singPath :: SPath ps

withIsPath2 :: forall {c :: k} (qs :: Path kk k1 c) r. IsPath qs => (Ob (ps +++ qs) => r) -> r

withPathO2 :: forall {c :: k} tag (qs :: Path kk k1 c) r. (IsOb tag ps, IsOb tag qs, Ob qs) => ((IsOb tag (ps +++ qs), Ob (ps +++ qs)) => r) -> r

withIsPath :: forall {k1} (kk :: CAT k1) (j :: k1) (k2 :: k1) (ps :: Path kk j k2) r. Bicategory kk => SPath ps -> ((Ob ps, Ob0 kk j, Ob0 kk k2) => r) -> r

append :: forall {k1} {kk :: CAT k1} {j :: k1} {b :: k1} {k2 :: k1} (ps :: Path kk j b) (qs :: Path kk b k2). SPath ps -> SPath qs -> SPath (ps +++ qs)

singleton :: forall {k} (kk :: CAT k) (i :: k) (j :: k) (p :: kk i j) (q :: kk i j). Bicategory kk => (p ~> q) -> (p '::: ('Nil :: Path kk j j)) ~> (q '::: ('Nil :: Path kk j j))

obj1 :: forall {k} {kk :: CAT k} {i :: k} {j :: k} (p :: kk i j). (Bicategory kk, Ob p, Ob0 kk i, Ob0 kk j) => Obj (p '::: ('Nil :: Path kk j j))

asObj :: forall {k} (kk :: CAT k) (i :: k) (j :: k) (ps :: Path kk i j). Bicategory kk => SPath ps -> Obj ps

concatFold :: forall {k1} {kk :: CAT k1} {i :: k1} {j :: k1} {k2 :: k1} (as :: Path kk i j) (bs :: Path kk j k2). Bicategory kk => SPath as -> SPath bs -> O (Fold bs) (Fold as) ~> Fold (as +++ bs)

splitFold :: forall {k1} {kk :: CAT k1} {i :: k1} {j :: k1} {k2 :: k1} (as :: Path kk i j) (bs :: Path kk j k2). Bicategory kk => SPath as -> SPath bs -> Fold (as +++ bs) ~> O (Fold bs) (Fold as)

type family Fold (as :: Path kk j k1) :: kk j k1

fold :: forall {k} {kk :: CAT k} {i :: k} {j :: k} (as :: Path kk i j). Bicategory kk => SPath as -> Fold as ~> Fold as

combineAll :: forall {k} {kk :: CAT k} {i :: k} {j :: k} (as :: Path kk i j). (Bicategory kk, Ob as) => as ~> (Fold as '::: ('Nil :: Path kk j j))

splitAll :: forall {k} {kk :: CAT k} {i :: k} {j :: k} (as :: Path kk i j). (Bicategory kk, Ob as) => (Fold as '::: ('Nil :: Path kk j j)) ~> as

data Strictified (ps :: Path kk j k1) (qs :: Path kk j k1) where

Constructors

St :: forall {k} {kk :: CAT k} {j :: k} {k1 :: k} (ps :: Path kk j k1) (qs :: Path kk j k1). (Ob ps, Ob qs) => (Fold ps ~> Fold qs) -> Strictified ps qs

view :: forall {k1} (kk :: CAT k1) (j :: k1) (k2 :: k1) (ps :: Path kk j k2) (qs :: Path kk j k2). Strictified ps qs -> (Strictified ps qs, SPath ps, SPath qs)

pattern Str :: Bicategory kk => (Ob0 kk j, Ob0 kk k2) => SPath ps -> SPath qs -> (Fold ps ~> Fold qs) -> Strictified ps qs

unStr :: forall {k1} {kk :: CAT k1} {j :: k1} {k2 :: k1} (ps :: Path kk j k2) (qs :: Path kk j k2). Strictified ps qs -> Fold ps ~> Fold qs

introI :: forall {k} {kk :: CAT k} {j :: k}. (Ob0 kk j, Bicategory kk) => ('Nil :: Path kk j j) ~> ((I :: kk j j) '::: ('Nil :: Path kk j j))

elimI :: forall {s} {kk :: CAT s} {j :: s}. (Ob0 kk j, Bicategory kk) => ((I :: kk j j) '::: ('Nil :: Path kk j j)) ~> ('Nil :: Path kk j j)

introO :: forall {k1} {kk :: k1 -> k1 -> Type} {i :: k1} {j :: k1} {k2 :: k1} (p :: kk i j) (q :: kk j k2). (Ob0 kk i, Ob0 kk j, Ob0 kk k2, Bicategory kk, Ob p, Ob q) => (p '::: (q '::: ('Nil :: Path kk k2 k2))) ~> (O q p '::: ('Nil :: Path kk k2 k2))

elimO :: forall {k1} {kk :: k1 -> k1 -> Type} {i :: k1} {j :: k1} {k2 :: k1} (p :: kk i j) (q :: kk j k2). (Ob0 kk i, Ob0 kk j, Ob0 kk k2, Bicategory kk, Ob p, Ob q) => (O q p '::: ('Nil :: Path kk k2 k2)) ~> (p '::: (q '::: ('Nil :: Path kk k2 k2)))

stUnit :: forall {k} {kk :: k -> k -> Type} {i :: k} {j :: k} (p :: kk i j) (q :: kk j i). (Adjunction p q, Ob p, Ob q) => ('Nil :: Path kk i i) ~> (p '::: (q '::: ('Nil :: Path kk i i)))

stCounit :: forall {k} {kk :: k -> k -> Type} {i :: k} {j :: k} (p :: kk i j) (q :: kk j i). (Adjunction p q, Ob p, Ob q) => (q '::: (p '::: ('Nil :: Path kk j j))) ~> ('Nil :: Path kk j j)

withIsObTagFold :: forall {k1} {kk :: CAT k1} {j :: k1} {k2 :: k1} tag (ps :: Path kk j k2) r. (Bicategory (SUBCAT tag kk), WithObO2 tag kk, IsOb tag ps, Ob ps) => ((IsOb tag (Fold ps), Ob (Fold ps)) => r) -> r