Proarrow.Category.Enriched

Documentation

class (Monoidal v, Profunctor p) => Enriched v (p :: j +-> k) where

Working with enriched categories and profunctors in Haskell is hard. Instead we encode them using the underlying regular category/profunctor, and show that the enriched structure can be recovered.

Associated Types

type Hom v (p :: j +-> k) (a :: k) (b :: j) :: v

Methods

underlying :: forall (a :: k) (b :: j). p a b -> (Unit :: v) ~> Hom v p a b

enriched :: forall (a :: k) (b :: j). (Ob a, Ob b) => ((Unit :: v) ~> Hom v p a b) -> p a b

type Self k = SUBCAT (Any :: k -> Constraint)