Proarrow.Category.Equipment

Documentation

class (Bicategory hk, Bicategory vk, forall (k :: c). Ob0 vk k => Ob0' hk k) => HasCompanions (hk :: CAT c) (vk :: CAT c) | hk -> vk where

A double category with companions.

Associated Types

type Companion (hk :: CAT c) (f :: vk j k) :: hk j k

Methods

mapCompanion :: forall {j :: c} {k :: c} (f :: vk j k) (g :: vk j k). (f ~> g) -> Companion hk f ~> Companion hk g

compToId :: forall (k :: c). Ob0 vk k => Companion hk (I :: vk k k) ~> (I :: hk k k)

compFromId :: forall (k :: c). Ob0 vk k => (I :: hk k k) ~> Companion hk (I :: vk k k)

compToCompose :: forall {i :: c} {j :: c} {k :: c} (f :: vk j k) (g :: vk i j). Obj f -> Obj g -> Companion hk (O f g) ~> O (Companion hk f) (Companion hk g)

compFromCompose :: forall {j1 :: c} {j2 :: c} {k :: c} (f :: vk j2 k) (g :: vk j1 j2). Obj f -> Obj g -> O (Companion hk f) (Companion hk g) ~> Companion hk (O f g)

type SQ' (hk :: CAT c) (vk :: CAT c) (h :: c) (i :: c) (j :: c) (k :: c) = PRO (hk h j, vk j k) (hk i k, vk h i)

The kind of a square '(p, g) '(q, f).

h--f--i
|  v  |
p--@--q
|  v  |
j--g--k

type SQ (hk :: CAT c) (vk :: CAT c) = forall {h :: c} {i :: c} {j :: c} {k :: c}. SQ' hk vk h i j k

data Sq (pf :: (hk h j, vk j k)) (qg :: (hk i k, vk h i)) where

Constructors

Sq :: forall {c} {hk :: CAT c} {vk :: CAT c} {h :: c} {i :: c} {j :: c} {k :: c} (p :: hk h j) (q :: hk i k) (f :: vk h i) (g :: vk j k). (Ob0 vk h, Ob0 vk i, Ob0 vk j, Ob0 vk k, Ob p, Ob q, Ob f, Ob g) => (O (Companion hk g) p ~> O q (Companion hk f)) -> Sq '(p, g) '(q, f)

object :: forall {s} hk vk (k :: s). (HasCompanions hk vk, Ob0 vk k) => Sq '(I :: hk k k, I :: vk k k) '(I :: hk k k, I :: vk k k)

The empty square for an object.

hArr :: forall {s} {hk} {vk} {j :: s} {k :: s} (p :: hk j k) (q :: hk j k). (HasCompanions hk vk, Ob0 vk j, Ob0 vk k) => (p ~> q) -> Sq '(p, I :: vk k k) '(q, I :: vk j j)

Make a square from a horizontal arrow

hId :: forall {s} hk vk (j :: s) (k :: s) (p :: hk j k). (HasCompanions hk vk, Ob0 vk j, Ob0 vk k, Ob p) => Sq '(p, I :: vk k k) '(p, I :: vk j j)

compId :: forall {s} {hk} {vk} {j :: s} {k :: s} (f :: vk j k). (HasCompanions hk vk, Ob0 vk j, Ob0 vk k, Ob f) => Sq '(Companion hk f, I :: vk k k) '(Companion hk f, I :: vk j j)

(|||) :: forall {s} {k1 :: s} {hk} {vk} {h :: s} {i :: s} {j :: s} {k2 :: s} {l :: s} (p :: hk h j) (q :: hk i k2) (r :: hk l k1) (f :: vk h i) (g :: vk j k2) (d :: vk i l) (e :: vk k2 k1). HasCompanions hk vk => Sq '(p, g) '(q, f) -> Sq '(q, e) '(r, d) -> Sq '(p, O e g) '(r, O d f) infixl 6

Horizontal composition

vArr :: forall {s} {hk} {vk} {j :: s} {k :: s} (f :: vk j k) (g :: vk j k). (HasCompanions hk vk, Ob0 vk j, Ob0 vk k) => (f ~> g) -> Sq '(I :: hk j j, f) '(I :: hk k k, g)

Make a square from a vertical arrow

vId :: forall {s} hk vk (j :: s) (k :: s) (f :: vk j k). (HasCompanions hk vk, Ob0 vk j, Ob0 vk k, Ob f) => Sq '(I :: hk j j, f) '(I :: hk k k, f)

(===) :: forall {c} {k1 :: c} {hk} {vk} {h :: c} {i :: c} {j :: c} {k2 :: c} {l :: c} (p :: hk h j) (q :: hk i l) (r :: hk j k2) (s :: hk l k1) (f :: vk h i) (g :: vk j l) (e :: vk k2 k1). HasCompanions hk vk => Sq '(p, g) '(q, f) -> Sq '(r, e) '(s, g) -> Sq '(O r p, e) '(O s q, f) infixl 5

Vertical composition

fromRight :: forall {s} {hk} {vk} {j :: s} {k :: s} (f :: vk j k). (HasCompanions hk vk, Ob' f) => Sq '(I :: hk j j, f) '(Companion hk f, I :: vk j j)

toLeft :: forall {s} {hk} {vk} {j :: s} {k :: s} (f :: vk j k). (HasCompanions hk vk, Ob' f) => Sq '(Companion hk f, I :: vk k k) '(I :: hk k k, f)

toRight :: forall {c} {hk} {vk} {j :: c} {k :: c} (f :: vk j k). (Equipment hk vk, Ob0 vk j, Ob0 vk k, Ob f) => Sq '(I :: hk j j, I :: vk j j) '(Conjoint hk f, f)

fromLeft :: forall {s} {hk} {vk} {j :: s} {k :: s} (f :: vk j k). (Equipment hk vk, Ob0 vk j, Ob0 vk k, Ob f) => Sq '(Conjoint hk f, f) '(I :: hk k k, I :: vk k k)

flipCompanion :: forall {s} {i :: s} {j :: s} {k :: s} (hk :: CAT s) (vk :: s -> s -> Type) (f :: vk j k) (p :: hk i j) (q :: hk i k). (Equipment hk vk, Ob p) => Obj f -> (O (Companion hk f) p ~> q) -> p ~> O (Conjoint hk f) q

flipCompanionInv :: forall {s} {i :: s} {j :: s} {k :: s} (hk :: CAT s) (vk :: s -> s -> Type) (f :: vk j k) (p :: hk i j) (q :: hk i k). (Equipment hk vk, Ob q) => Obj f -> (p ~> O (Conjoint hk f) q) -> O (Companion hk f) p ~> q

flipConjoint :: forall {c} {k1 :: c} {j :: c} {k2 :: c} (hk :: CAT c) (vk :: c -> c -> Type) (f :: vk j k2) (p :: hk j k1) (q :: hk k2 k1). (Equipment hk vk, Ob p) => Obj f -> (O p (Conjoint hk f) ~> q) -> p ~> O q (Companion hk f)

flipConjointInv :: forall {c} {k1 :: c} {j :: c} {k2 :: c} (hk :: CAT c) (vk :: c -> c -> Type) (f :: vk j k2) (p :: hk j k1) (q :: hk k2 k1). (Equipment hk vk, Ob q) => Obj f -> (p ~> O q (Companion hk f)) -> O p (Conjoint hk f) ~> q

class Adjunction (Companion hk f) (Conjoint hk f) => ComConAdjunction (hk :: CAT c) (vk1 :: k) (f :: vk j k1)

class HasCompanions hk vk => Equipment (hk :: CAT c) (vk :: CAT c) | hk -> vk where

Minimal complete definition

mapConjoint, (conjToId, conjFromId, conjToCompose, conjFromCompose | comConUnit, comConCounit)

Associated Types

type Conjoint (hk :: CAT c) (f :: vk j k) :: hk k j

Methods

mapConjoint :: forall {j :: c} {k :: c} (f :: vk j k) (g :: vk j k). (f ~> g) -> Conjoint hk g ~> Conjoint hk f

conjToId :: forall (k :: c). Ob0 vk k => Conjoint hk (I :: vk k k) ~> (I :: hk k k)

conjFromId :: forall (k :: c). Ob0 vk k => (I :: hk k k) ~> Conjoint hk (I :: vk k k)

conjToCompose :: forall {k1 :: c} {j :: c} {k2 :: c} (f :: vk j k2) (g :: vk k1 j). Obj f -> Obj g -> Conjoint hk (O f g) ~> O (Conjoint hk g) (Conjoint hk f)

conjFromCompose :: forall {j1 :: c} {j2 :: c} {k :: c} (f :: vk j2 k) (g :: vk j1 j2). Obj f -> Obj g -> O (Conjoint hk g) (Conjoint hk f) ~> Conjoint hk (O f g)

comConUnit :: forall {j :: c} {k :: c} (f :: vk j k). Obj f -> (I :: hk j j) ~> O (Conjoint hk f) (Companion hk f)

default comConUnit :: forall {j :: c} {k :: c} (f :: vk j k). (Ob f => ComConAdjunction hk vk f) => Obj f -> (I :: hk j j) ~> O (Conjoint hk f) (Companion hk f)

comConCounit :: forall {j :: c} {k :: c} (f :: vk j k). Obj f -> O (Companion hk f) (Conjoint hk f) ~> (I :: hk k k)

default comConCounit :: forall {j :: c} {k :: c} (f :: vk j k). (Ob f => ComConAdjunction hk vk f) => Obj f -> O (Companion hk f) (Conjoint hk f) ~> (I :: hk k k)

type Cart (p :: hk b d) (f :: vk a b) (g :: vk c1 d) = O (O (Conjoint hk g) p) (Companion hk f)

P(f, g)

data RetroSq (pf :: (hk i k, vk h i)) (qg :: (hk h j, vk j k)) where

The kind of a retro square '(q, f) '(p, g).

h--f--i
|  v  |
p----q
|  v  |
j--g--k

Constructors

RetroSq :: forall {c} {hk :: CAT c} {vk :: CAT c} {h :: c} {i :: c} {j :: c} {k :: c} (p :: hk h j) (q :: hk i k) (f :: vk h i) (g :: vk j k). (Ob0 vk h, Ob0 vk i, Ob0 vk j, Ob0 vk k, Ob p, Ob q, Ob f, Ob g) => (O q (Companion hk f) ~> O (Companion hk g) p) -> RetroSq '(q, f) '(p, g)

companionFold :: forall {k1} {hk :: CAT k1} {vk :: CAT k1} {j :: k1} {k2 :: k1} (fs :: Path vk j k2). HasCompanions hk vk => SPath fs -> Companion hk (Fold fs) ~> Fold (Companion (Path hk) fs)

foldCompanion :: forall {k1} {hk :: CAT k1} {vk :: CAT k1} {j :: k1} {k2 :: k1} (fs :: Path vk j k2). HasCompanions hk vk => SPath fs -> Fold (Companion (Path hk) fs) ~> Companion hk (Fold fs)

mapCompanionSPath :: forall {k1} (hk :: CAT k1) {vk :: CAT k1} {j :: k1} {k2 :: k1} (fs :: Path vk j k2). HasCompanions hk vk => SPath fs -> SPath (Companion (Path hk) fs)

mapConjointSPath :: forall {k1} (hk :: CAT k1) {vk :: CAT k1} {j :: k1} {k2 :: k1} (fs :: Path vk j k2). Equipment hk vk => SPath fs -> SPath (Conjoint (Path hk) fs)

conjId :: forall {k1} {hk :: CAT k1} {vk :: CAT k1} {j :: k1} {k2 :: k1} (f :: vk j k2). (Equipment hk vk, Ob0 vk j, Ob0 vk k2, Ob f) => Sq '(Conjoint hk f '::: ('Nil :: Path hk j j), 'Nil :: Path vk j j) '(Conjoint hk f '::: ('Nil :: Path hk j j), 'Nil :: Path vk k2 k2)

fromLeft1 :: forall {k1} (hk :: CAT k1) (vk :: CAT k1) (j :: k1) (k2 :: k1) (f :: vk j k2). (Equipment hk vk, Ob' f) => Sq '(Conjoint hk f '::: ('Nil :: Path hk j j), f '::: ('Nil :: Path vk k2 k2)) '('Nil :: Path hk k2 k2, 'Nil :: Path vk k2 k2)

fromLeft2 :: forall {k1} {hk :: CAT k1} {vk :: k1 -> k1 -> Type} {i :: k1} {j :: k1} {k2 :: k1} (f :: vk i j) (g :: vk j k2). (Equipment hk vk, Ob0 vk i, Ob0 vk j, Ob0 vk k2, Ob f, Ob g) => Sq '(Conjoint hk g '::: (Conjoint hk f '::: ('Nil :: Path hk i i)), f '::: (g '::: ('Nil :: Path vk k2 k2))) '('Nil :: Path hk k2 k2, 'Nil :: Path vk k2 k2)

adjVK :: forall {k1} {j1 :: k1} hk vk (i :: k1) (j2 :: k1) (k2 :: k1) (f :: vk j2 k2) (g :: vk j1 i) (v :: hk j2 j1) (w :: hk k2 i) (x :: hk j1 j2) (y :: hk i k2). (Adjunction x v, Adjunction y w, HasCompanions hk vk, Ob v, Ob w) => RetroSq '(y, g) '(x, f) -> Sq '(v, g) '(w, f)