Proarrow.Category.Equipment.Stateful

Documentation

newtype STT m k (i :: ()) (j :: ())

Constructors

ST (k +-> k)

data StT (a :: STT m k i j) (b :: STT m k i j) where

Constructors

StT :: forall m k (p :: k +-> k) (q :: k +-> k) (i :: ()) (j :: ()). (Strong m p, Strong m q) => (p :~> q) -> StT ('ST p :: STT m k i j) ('ST q :: STT m k i j)

type STSq (p :: k +-> k) (q :: k +-> k) (a :: m) (b :: m) = Sq '('ST p :: STT m k '() '(), 'MK a :: MonK m '() '()) '('ST q :: STT m k '() '(), 'MK b :: MonK m '() '())

unSTSq :: forall {m} {k} p q (a :: m) (b :: m). SymMonoidal m => STSq p q a b -> (forall (x :: k) (y :: k). p x y -> q (Act a x) (Act b y), Obj a, Obj b, Obj ('ST p :: STT m k '() '()), Obj ('ST q :: STT m k '() '()))

pattern STSq :: forall {m} {k} p q (a :: m) (b :: m). (SymMonoidal m, MonoidalAction m k) => (Strong m p, Strong m q, Ob a, Ob b) => (forall (x :: k) (y :: k). p x y -> q (Act a x) (Act b y)) -> STSq p q a b

crossing :: forall {k} {m} (p :: k +-> k) (a :: m). (Strong m p, Ob a, SymMonoidal m) => STSq p p a a

pi0 :: forall {m} {k} (p :: k +-> k) (q :: k +-> k). (Strong m p, Strong m q, SymMonoidal m) => STSq (p :*: q) p (Unit :: m) (Unit :: m)

pi1 :: forall {m} {k} (p :: k +-> k) (q :: k +-> k). (Strong m p, Strong m q, SymMonoidal m) => STSq (p :*: q) q (Unit :: m) (Unit :: m)

mult :: forall {m} {k} (p :: k +-> k) (q :: k +-> k) (r :: k +-> k) (a :: m) (b :: m). (Strong m p, Strong m q, Strong m r, SymMonoidal m, Ob a, Ob b) => STSq p q a b -> STSq p r a b -> STSq p (q :*: r) a b

ip0 :: forall {m} {k} (p :: k +-> k) (q :: k +-> k). (Strong m p, Strong m q, SymMonoidal m) => STSq p (p :+: q) (Unit :: m) (Unit :: m)

ip1 :: forall {m} {k} (p :: k +-> k) (q :: k +-> k). (Strong m p, Strong m q, SymMonoidal m) => STSq q (p :+: q) (Unit :: m) (Unit :: m)

plus :: forall {m} {k} (p :: k +-> k) (q :: k +-> k) (r :: k +-> k) (a :: m) (b :: m). (Strong m p, Strong m q, Strong m r, SymMonoidal m, Ob a, Ob b) => STSq p r a b -> STSq q r a b -> STSq (p :+: q) r a b

sigma0 :: forall {m} {k} (a :: m) (b :: m). (Ob a, Ob b, SymMonoidal m, MonoidalAction m k, HasCoproducts m) => STSq (Id :: k -> k -> Type) (Id :: k -> k -> Type) a (a || b)

sigma1 :: forall {m} {k} (a :: m) (b :: m). (Ob a, Ob b, SymMonoidal m, MonoidalAction m k, HasCoproducts m) => STSq (Id :: k -> k -> Type) (Id :: k -> k -> Type) b (a || b)

copair :: forall {k} (p :: k +-> k) (q :: k +-> k) (a :: k) (b :: k) (c :: k). (Strong k p, Strong k q, SymMonoidal k, Distributive k, SelfAction k, MonoidalProfunctor (Coprod q), Ob a, Ob b, Ob c) => STSq p q a c -> STSq p q b c -> STSq p q (a || b) c