Proarrow.Category.Instance.Free

Documentation

newtype FREE (cs :: [Kind -> Constraint]) (p :: CAT j) k

Constructors

EMB j

type family All (cs :: [Kind -> Constraint]) k where ...

Equations

All ('[] :: [Kind -> Constraint]) k = ()

All (c ': cs) k = (c k, All cs k)

class Includes (cs :: k) (c :: Kind -> Constraint)

data Free (a :: FREE cs p k) (b :: FREE cs p k) where

Constructors

Id :: forall {j} {cs :: [Kind -> Constraint]} {p :: CAT j} {k} (a :: FREE cs p k). Ob a => Free a a

Emb :: forall {j} (a1 :: j) (b1 :: j) (p :: CAT j) (cs :: [Kind -> Constraint]) k (a :: FREE cs p k). (Ob a1, Ob b1, Typeable a1, Typeable b1) => p a1 b1 -> Free a ('EMB a1 :: FREE cs p k) -> Free a ('EMB b1 :: FREE cs p k)

Str :: forall {j} {k} {cs :: [Kind -> Constraint]} {p :: CAT j} (c :: Kind -> Constraint) (a1 :: FREE cs p k) (b :: FREE cs p k) (a :: FREE cs p k). (All cs k => c k, Includes cs c, HasStructure k cs p c, Ob a1, Ob b) => Struct c a1 b -> Free a a1 -> Free a b

emb :: forall {j} (a :: j) (b :: j) (cs :: [Kind -> Constraint]) p k. (Ob a, Ob b, Typeable a, Typeable b, Ok cs p k) => p a b %1 -> Free ('EMB a :: FREE cs p k) ('EMB b :: FREE cs p k)

class (forall (x :: k) (y :: k1). Eq (p x y)) => Eq2 (p :: k -> k1 -> Type)

class (forall (x :: k) (y :: k1). Show (p x y)) => Show2 (p :: k -> k1 -> Type)

class (Typeable p, Typeable c, Typeable k2, Typeable j) => Ok (c :: k) (p :: CAT j) (k2 :: k1)

class (Ok c p k, Eq2 p) => WithEq (a :: FREE c p k)

class IsFreeOb (a :: FREE cs p k) where

Associated Types

type Lower (f :: j +-> k) (a :: FREE cs p k) :: k

Methods

withLowerOb :: forall (f :: j +-> k) r. Representable f => (Ob (Lower f a) => r) -> r

class ((Ok c p k, Eq2 p) => Eq2 str, Ok c p k => Typeable str) => CanEq (str :: CAT (FREE c p k))

class (Typeable c, CanEq (Struct c :: CAT (FREE cs p k))) => HasStructure k (cs :: [Kind -> Constraint]) (p :: CAT j) (c :: Kind -> Constraint) where

Associated Types

data Struct (c :: Kind -> Constraint) :: CAT (FREE cs p k)

Methods

foldStructure :: forall (f :: j +-> k) (a :: FREE cs p k) (b :: FREE cs p k). (c k, Representable f) => (forall (x :: FREE cs p k) (y :: FREE cs p k). (x ~> y) -> Lower f x ~> Lower f y) -> Struct c a b -> Lower f a ~> Lower f b

fold :: forall {j} {k} {p} (cs :: [Kind -> Constraint]) f (a :: FREE cs p k) (b :: FREE cs p k). (All cs k, Representable f) => (forall (x :: j) (y :: j). p x y -> f (f % x) y) -> (a ~> b) -> Lower f a ~> Lower f b

retract :: forall {j} {k} (cs :: [Kind -> Constraint]) (f :: j +-> k) (a :: FREE cs (InitialProfunctor :: j -> j -> Type) k) (b :: FREE cs (InitialProfunctor :: j -> j -> Type) k). (All cs k, Representable f) => (a ~> b) -> Lower f a ~> Lower f b

type Req (c :: Type -> Constraint) (lut :: [(Symbol, k)]) (a :: FREE cs p k) (b :: FREE cs p k) = ()

data AssertEqs (c :: Kind -> Constraint) k where

Constructors

NoLaws :: forall (c :: Kind -> Constraint) k. AssertEqs c k

(:>>:) :: forall (c :: Kind -> Constraint) k. AssertEqs c k -> AssertEqs c k -> AssertEqs c k

(:=:) :: forall {k} {c :: Kind -> Constraint} (a :: FREE '[c] (Var c) k) (b :: FREE '[c] (Var c) k). (forall r (lut :: [(Symbol, k)]). ((Req Eq lut a b, Req Show lut a b) => Free a b -> Free a b -> r) -> r) -> AssertEqs c k infix 0

(>>) :: forall (c :: Kind -> Constraint) k. AssertEqs c k -> AssertEqs c k -> AssertEqs c k

class Laws (c :: Kind -> Constraint) where

Associated Types

data Var (c :: Kind -> Constraint) (a :: Symbol) (b :: Symbol)

Methods

laws :: (c k, Typeable k) => AssertEqs c k

type family AssocLookup (lut :: [(Symbol, k)]) (a :: Symbol) :: k where ...

Equations

AssocLookup ('[] :: [(Symbol, k)]) a = Any :: k

AssocLookup ('(s, k2) ': lut :: [(Symbol, k1)]) s = k2

AssocLookup ('(s, k2) ': lut :: [(Symbol, k1)]) a = AssocLookup lut a

type family AllOb (lut :: [(Symbol, k)]) where ...

Equations

AllOb ('[] :: [(Symbol, k)]) = ()

AllOb ('(s, k2) ': lut :: [(Symbol, k1)]) = (Ob k2, AllOb lut)

data Lookup (lut :: [(Symbol, k)]) (a :: k) (b :: Symbol) where

Constructors

Lookup :: forall {k} (lut :: [(Symbol, k)]) (b :: Symbol) (a :: k). Ob b => (a ~> (Lookup lut % b)) -> Lookup lut a b

data family TermF :: k

data family InitF :: k

data family (a :: k) *! (b :: k) :: k

data family (a :: k) + (b :: k) :: k

data family UnitF :: k

data family (a :: k) **! (b :: k) :: k

data family (a :: k) --> (b :: k) :: k

data TestTy

Constructors

IntTy'

StringTy'

type IntTy = 'D 'IntTy'

type StringTy = 'D 'StringTy'

data Test (a :: DISCRETE TestTy) (b :: DISCRETE TestTy) where

Constructors

Show :: Test ('D 'IntTy') ('D 'StringTy')

Read :: Test ('D 'StringTy') ('D 'IntTy')

Succ :: Test ('D 'IntTy') ('D 'IntTy')

Dup :: Test ('D 'StringTy') ('D 'StringTy')

show :: forall (cs :: [Kind -> Constraint]) k (i :: FREE cs Test k). (i ~> ('EMB IntTy :: FREE cs Test k)) %1 -> i ~> ('EMB StringTy :: FREE cs Test k)

read :: forall (cs :: [Kind -> Constraint]) k (i :: FREE cs Test k). (i ~> ('EMB StringTy :: FREE cs Test k)) %1 -> i ~> ('EMB IntTy :: FREE cs Test k)

succ :: forall (cs :: [Kind -> Constraint]) k (i :: FREE cs Test k). (i ~> ('EMB IntTy :: FREE cs Test k)) %1 -> i ~> ('EMB IntTy :: FREE cs Test k)

dup :: forall (cs :: [Kind -> Constraint]) k (i :: FREE cs Test k). (i ~> ('EMB StringTy :: FREE cs Test k)) %1 -> i ~> ('EMB StringTy :: FREE cs Test k)

test :: forall (cs :: [Kind -> Constraint]) k (i :: FREE cs Test k). (i ~> ('EMB StringTy :: FREE cs Test k)) %1 -> i ~> ('EMB StringTy :: FREE cs Test k)

test2 :: forall (cs :: [Kind -> Constraint]) k (i :: FREE cs Test k). HasBinaryProducts (FREE cs Test k) => (i ~> ('EMB StringTy :: FREE cs Test k)) -> i ~> (('EMB StringTy :: FREE cs Test k) *! ('EMB StringTy :: FREE cs Test k))

data Interp a (b :: DISCRETE TestTy) where

Constructors

Interp :: forall (b :: DISCRETE TestTy) a. Ob b => (a ~> (Interp % b)) -> Interp a b

testFold :: String -> (String, String)

Orphan instances

CategoryOf Symbol

Instance details

Associated Types

type (~>)

Instance details

Defined in Proarrow.Category.Instance.Free

type (~>) = (:~:) :: Symbol -> Symbol -> Type

type Ob (a :: Symbol)

Instance details

Defined in Proarrow.Category.Instance.Free

type Ob (a :: Symbol) = ()

Profunctor ((:~:) :: Symbol -> Symbol -> Type)

Instance details

Methods

dimap :: forall (c :: Symbol) (a :: Symbol) (b :: Symbol) (d :: Symbol). (c ~> a) -> (b ~> d) -> (a :~: b) -> c :~: d

(\\) :: forall (a :: Symbol) (b :: Symbol) r. ((Ob a, Ob b) => r) -> (a :~: b) -> r

Promonad ((:~:) :: Symbol -> Symbol -> Type)

Instance details

Methods

id :: forall (a :: Symbol). Ob a => a :~: a

(.) :: forall (b :: Symbol) (c :: Symbol) (a :: Symbol). (b :~: c) -> (a :~: b) -> a :~: c