Proarrow.Category.Instance.ZX

Documentation

newtype Bitstring (n :: Nat)

Constructors

BS Int

split :: forall (n :: Nat) (m :: Natural). KnownNat n => Bitstring (n + m) -> (Bitstring n, Bitstring m)

Split n + m bits into two parts: the lower n bits and the higher m bits.

combine :: forall (n :: Nat) (m :: Nat). KnownNat n => Bitstring n -> Bitstring m -> Bitstring (n + m)

Combine two bitstrings of lengths n and m into one bitstring with the n lower bits or m higher bits.

type SparseMatrix (o :: Nat) (i :: Nat) = Map (Bitstring o, Bitstring i) (Complex Double)

epsilon :: Double

isZero :: Complex Double -> Bool

filterSparse :: forall (o :: Nat) (i :: Nat). SparseMatrix o i -> SparseMatrix o i

transpose :: forall (o :: Nat) (i :: Nat). SparseMatrix o i -> SparseMatrix i o

mirror :: forall (n :: Nat). KnownNat n => Bitstring n -> Bitstring (n + n)

enumAll :: forall (n :: Nat). KnownNat n => [Bitstring n]

nat :: forall (n :: Nat). KnownNat n => Int

withPlusIsNat :: forall (a :: Nat) (b :: Nat) r. (KnownNat a, KnownNat b) => (KnownNat (a + b) => r) -> r

data ZX (i :: Nat) (o :: Nat) where

Constructors

ZX :: forall (i :: Nat) (o :: Nat). (KnownNat i, KnownNat o) => SparseMatrix o i -> ZX i o

zSpider :: forall (i :: Nat) (o :: Nat). (KnownNat i, KnownNat o) => Double -> ZX i o

xSpider :: forall (i :: Nat) (o :: Nat). (KnownNat i, KnownNat o) => Double -> ZX i o

cup :: forall (n :: Nat). KnownNat n => ZX 0 (n + n)

cap :: forall (n :: Nat). KnownNat n => ZX (n + n) 0

zCopy :: ZX 1 2

zDisc :: ZX 1 0

xCopy :: ZX 1 2

xDisc :: ZX 1 0

zeroState :: ZX 0 1

oneState :: ZX 0 1

plusState :: ZX 0 1

minusState :: ZX 0 1

not :: ZX 1 1

cnot :: ZX 2 2

Controlled NOT gate

ghzState :: ZX 0 3

Greenberger–Horne–Zeilinger state

hadamard :: forall (n :: Nat). KnownNat n => ZX n n

Orphan instances

Monoidal Nat
Addition of the number of qubits as monoidal tensor. This is the Kronecker product of the matrices.

Instance details

Associated Types

type Unit

Instance details

Defined in Proarrow.Category.Instance.ZX

type Unit = 0

type (p :: Natural) ** (q :: Natural)

Instance details

Defined in Proarrow.Category.Instance.ZX

type (p :: Natural) ** (q :: Natural) = p + q

Methods

withOb2 :: forall (a :: Nat) (b :: Nat) r. (Ob a, Ob b) => (Ob (a ** b) => r) -> r

leftUnitor :: forall (a :: Nat). Ob a => ((Unit :: Nat) ** a) ~> a

leftUnitorInv :: forall (a :: Nat). Ob a => a ~> ((Unit :: Nat) ** a)

rightUnitor :: forall (a :: Nat). Ob a => (a ** (Unit :: Nat)) ~> a

rightUnitorInv :: forall (a :: Nat). Ob a => a ~> (a ** (Unit :: Nat))

associator :: forall (a :: Nat) (b :: Nat) (c :: Nat). (Ob a, Ob b, Ob c) => ((a ** b) ** c) ~> (a ** (b ** c))

associatorInv :: forall (a :: Nat) (b :: Nat) (c :: Nat). (Ob a, Ob b, Ob c) => (a ** (b ** c)) ~> ((a ** b) ** c)

SymMonoidal Nat

Instance details

Methods

swap :: forall (a :: Nat) (b :: Nat). (Ob a, Ob b) => (a ** b) ~> (b ** a)

CategoryOf Nat
The category of qubits, to implement ZX calculus from quantum computing.

Instance details

Associated Types

type (~>)

Instance details

Defined in Proarrow.Category.Instance.ZX

type (~>) = ZX

type Ob (a :: Nat)

Instance details

Defined in Proarrow.Category.Instance.ZX

type Ob (a :: Nat) = KnownNat a

CompactClosed Nat

Instance details

Methods

distribDual :: forall (a :: Nat) (b :: Nat). (Ob a, Ob b) => Dual (a ** b) ~> (Dual a ** Dual b)

dualUnit :: Dual (Unit :: Nat) ~> (Unit :: Nat)

StarAutonomous Nat

Instance details

Associated Types

type Dual (x :: Nat)

Instance details

Defined in Proarrow.Category.Instance.ZX

type Dual (x :: Nat) = x

Methods

dual :: forall (a :: Nat) (b :: Nat). (a ~> b) -> Dual b ~> Dual a

dualInv :: forall (a :: Nat) (b :: Nat). (Ob a, Ob b) => (Dual a ~> Dual b) -> b ~> a

linDist :: forall (a :: Nat) (b :: Nat) (c :: Nat). (Ob a, Ob b, Ob c) => ((a ** b) ~> Dual c) -> a ~> Dual (b ** c)

linDistInv :: forall (a :: Nat) (b :: Nat) (c :: Nat). (Ob a, Ob b, Ob c) => (a ~> Dual (b ** c)) -> (a ** b) ~> Dual c

Closed Nat

Instance details

Associated Types

type (x :: Nat) ~~> (y :: Nat)

Instance details

Defined in Proarrow.Category.Instance.ZX

type (x :: Nat) ~~> (y :: Nat) = ExpSA x y

Methods

withObExp :: forall (a :: Nat) (b :: Nat) r. (Ob a, Ob b) => (Ob (a ~~> b) => r) -> r

curry :: forall (a :: Nat) (b :: Nat) (c :: Nat). (Ob a, Ob b) => ((a ** b) ~> c) -> a ~> (b ~~> c)

apply :: forall (a :: Nat) (b :: Nat). (Ob a, Ob b) => ((a ~~> b) ** a) ~> b

(^^^) :: forall (a :: Nat) (b :: Nat) (x :: Nat) (y :: Nat). (b ~> y) -> (x ~> a) -> (a ~~> b) ~> (x ~~> y)

HasInitialObject Nat

Instance details

Associated Types

type InitialObject

Instance details

Defined in Proarrow.Category.Instance.ZX

type InitialObject = 0

Methods

initiate :: forall (a :: Nat). Ob a => (InitialObject :: Nat) ~> a

HasTerminalObject Nat

Instance details

Associated Types

type TerminalObject

Instance details

Defined in Proarrow.Category.Instance.ZX

type TerminalObject = 0

Methods

terminate :: forall (a :: Nat). Ob a => a ~> (TerminalObject :: Nat)