Proarrow.Core

Documentation

type PRO j k = j -> k -> Type

type CAT k = PRO k k

type BI k = (k, k) -> k

type OB k = k -> Constraint

type Kind = Type

class Any (a :: k)

class Promonad ((~>) :: CAT k) => CategoryOf k

Associated Types

type (~>) :: CAT k infixr 0

type Ob (a :: k) :: Constraint

type Ob a = Any a

type IsCategoryOf k cat = (CategoryOf k, cat ~ (~>) @k, Promonad cat)

type (:~>) p q = forall a b. p a b -> q a b infixr 0

class (CategoryOf j, CategoryOf k) => Profunctor (p :: PRO j k) where

Minimal complete definition

dimap

Methods

dimap :: (c ~> a) -> (b ~> d) -> p a b -> p c d

(\\) :: ((Ob a, Ob b) => r) -> p a b -> r infixl 1

default (\\) :: (Ob a, Ob b) => ((Ob a, Ob b) => r) -> p a b -> r

(//) :: Profunctor p => p a b -> ((Ob a, Ob b) => r) -> r infixr 0

lmap :: Profunctor p => (c ~> a) -> p a b -> p c b

rmap :: Profunctor p => (b ~> d) -> p a b -> p a d

dimapDefault :: Promonad p => p c a -> p b d -> p a b -> p c d

class Profunctor p => Promonad p where

Methods

id :: Ob a => p a a

(.) :: p b c -> p a b -> p a c infixr 9

arr :: Promonad p => (a ~> b) -> p a b

type Obj a = a ~> a

obj :: forall {k} (a :: k). (CategoryOf k, Ob a) => Obj a

src :: forall {k} a b p. Profunctor p => p (a :: k) b -> Obj a

tgt :: forall {k} a b p. Profunctor p => p (a :: k) b -> Obj b

type family UN (w :: j -> k) (wa :: k) :: j

A helper type family to unwrap a wrapped kind. This is needed because the field selector functions of newtypes have to be lower case and therefore cannot be used at the type level.

type Is w a = a ~ w (UN w a)

Is w a checks that the kind a is a kind wrapped by w.