Proarrow.Profunctor.Representable

Documentation

class Profunctor p => Representable (p :: j +-> k) where

Associated Types

type (p :: j +-> k) % (a :: j) :: k infixl 8

Methods

index :: forall (a :: k) (b :: j). p a b -> a ~> (p % b)

tabulate :: forall (b :: j) (a :: k). Ob b => (a ~> (p % b)) -> p a b

repMap :: forall (a :: j) (b :: j). (a ~> b) -> (p % a) ~> (p % b)

repObj :: forall {j} {k} (p :: j +-> k) (a :: j). (Representable p, Ob a) => Obj (p % a)

withRepOb :: forall {j} {k} (p :: j +-> k) (a :: j) r. (Representable p, Ob a) => (Ob (p % a) => r) -> r

dimapRep :: forall {j} {k} p (a :: k) (b :: j) (c :: k) (d :: j). Representable p => (c ~> a) -> (b ~> d) -> p a b -> p c d

trivialRep :: forall {j} {k} p (a :: j). (Representable p, Ob a) => p (p % a) a

data CorepStar (p :: k +-> j) (a :: k) (b :: j) where

Constructors

CorepStar

Fields

:: forall {j} {k} (b :: j) (a :: k) (p :: k +-> j). Ob b

=> { unCorepStar :: a ~> (p %% b)

} -> CorepStar p a b

data RepCostar (p :: k +-> j) (a :: k) (b :: j) where

Constructors

RepCostar

Fields

:: forall {k} {j} (a :: k) (p :: k +-> j) (b :: j). Ob a

=> { unRepCostar :: (p % a) ~> b

} -> RepCostar p a b

flipRep :: forall {k1} {k} (p :: k1 +-> k) (q :: k +-> k1). (Representable p, Corepresentable q) => (RepCostar p :~> q) -> CorepStar q :~> p

class (CategoryOf j, CategoryOf k) => FunctorForRep (f :: j +-> k) where

A perfectly valid functor definition, but hard to use. So we only use it to easily make representable profunctors with Rep.

Associated Types

type (f :: j +-> k) @ (a :: j) :: k

Methods

fmap :: forall (a :: j) (b :: j). (a ~> b) -> (f @ a) ~> (f @ b)

data Rep (f :: j +-> k) (a :: k) (b :: j) where

Constructors

Rep

Fields

:: forall {j} {k} (f :: j +-> k) (a :: k) (b :: j). Ob b

=> { unRep :: a ~> (f @ b)

} -> Rep f a b