Proarrow.Squares.Limit

Documentation

limit :: forall {k1} {hk :: k1 -> k1 -> Type} {vk :: k1 -> k1 -> Type} {a :: k1} {i :: k1} {k2 :: k1} (j :: hk i a) (d :: vk i k2). (Ob0 vk i, HasLimits vk j k2, Ob d) => Sq '(j '::: ('Nil :: Path hk a a), Limit j d '::: ('Nil :: Path vk k2 k2)) '('Nil :: Path hk k2 k2, d '::: ('Nil :: Path vk k2 k2))

The projection out of the j-weighted limit l of d.

a--l--k
|  v  |
j--@  |
|  v  |
i--d--k

limitUniv :: forall {k1} {hk :: k1 -> k1 -> Type} {vk :: k1 -> k1 -> Type} {a :: k1} {i :: k1} {k2 :: k1} (j :: hk i a) (d :: vk i k2) (p :: vk a k2). (Ob0 vk i, HasLimits vk j k2, Ob d, Ob p) => Sq '(j '::: ('Nil :: Path hk a a), p '::: ('Nil :: Path vk k2 k2)) '('Nil :: Path hk k2 k2, d '::: ('Nil :: Path vk k2 k2)) -> Sq '('Nil :: Path hk a a, p '::: ('Nil :: Path vk k2 k2)) '('Nil :: Path hk k2 k2, Limit j d '::: ('Nil :: Path vk k2 k2))

The universal property of the limit.

The dotted inner square is the input square, the outer square is the output square.

a-+--p--+-k
| :  v  : |
| j--@  : |
| :  v  : |
| i..d..k |
a----l----k

rightAdjointPreservesLimits :: forall {k1} {hk :: k1 -> k1 -> Type} {vk :: k1 -> k1 -> Type} {k2 :: k1} {k' :: k1} {i :: k1} {a :: k1} (f :: vk k' k2) (g :: vk k2 k') (d :: vk i k2) (j :: hk i a). (Adjunction f g, HasCompanions hk vk, Ob d, Ob f, Ob g, HasLimits vk j k2, HasLimits vk j k', Ob0 vk i, Ob0 vk k2, Ob0 vk k', Ob0 vk a) => Sq '('Nil :: Path hk a a, Limit j (O g d) '::: ('Nil :: Path vk k' k')) '('Nil :: Path hk k' k', Limit j d '::: (g '::: ('Nil :: Path vk k' k')))

Preservation of limits by right adjoints

Let l be the j-weighted limit of d and l' be the j-weighted limit of g ∘ d. Then we provide the following square:

a--l'-k'
|  v  |
|  @  |
| / \ |
| v v |
a-l-g-k'

With this implementation:

+-----l'--+---------+
|     v   |   /@\   |
|     |   |  v   v  |
+-+---l'--+-f-+-+-g-+
| :   v   | v : | v |
| j---@   | | : | | |
| :  / \  | | : | | |
| : v   v | v : | | |
| +-d-+-g-+-f-+ | | |
| : | | v   v : | | |
| : v |  \@/  : | | |
| +.d.+.......+ | v |
+-------l-------+-g-+

rightAdjointPreservesLimitsInv :: forall {k1} {hk :: k1 -> k1 -> Type} {vk :: k1 -> k1 -> Type} {k2 :: k1} {k' :: k1} {i :: k1} {a :: k1} (g :: vk k2 k') (d :: vk i k2) (j :: hk i a). (HasCompanions hk vk, Ob d, Ob g, HasLimits vk j k2, HasLimits vk j k', Ob0 vk i, Ob0 vk k2, Ob0 vk k', Ob0 vk a) => Sq '('Nil :: Path hk a a, Limit j d '::: (g '::: ('Nil :: Path vk k' k'))) '('Nil :: Path hk k' k', Limit j (O g d) '::: ('Nil :: Path vk k' k'))

The inverse works for any arrow:

The required square

a-l-g-k'
| v v |
| \@/ |
|  |  |
a--l'-k'

is implemented as

a-+-l-+-g-+-k'
| : v | v : |
| j-@ | | : |
| : v | v : |
| +.d.+.g.+ |
a-----l'----k'

unit :: forall {k1} {hk :: CAT k1} {vk :: k1 -> k1 -> Type} {j :: k1} {k2 :: k1} (f :: vk j k2) (g :: vk k2 j). (Adjunction f g, HasCompanions hk vk, Ob f, Ob g, Ob0 vk j, Ob0 vk k2) => Sq '('Nil :: Path hk j j, 'Nil :: Path vk j j) '('Nil :: Path hk j j, f '::: (g '::: ('Nil :: Path vk j j)))

counit :: forall {k1} {hk :: CAT k1} {vk :: k1 -> k1 -> Type} {j :: k1} {k2 :: k1} (f :: vk j k2) (g :: vk k2 j). (Adjunction f g, HasCompanions hk vk, Ob f, Ob g, Ob0 vk j, Ob0 vk k2) => Sq '('Nil :: Path hk k2 k2, g '::: (f '::: ('Nil :: Path vk k2 k2))) '('Nil :: Path hk k2 k2, 'Nil :: Path vk k2 k2)