Proarrow.Universal

Documentation

class (Representable r, Ob a) => InitUniversal (r :: j +-> k) (a :: k) where

The initial universal property of a functor r (as a representable profunctor) and an object a.

Associated Types

type L (r :: j +-> k) (a :: k) :: j

Methods

initUnivArr :: r a (L r a)

initUnivProp :: forall (b :: j). r a b -> L r a ~> b

class (Corepresentable l, Ob b) => TermUniversal (l :: j +-> k) (b :: j) where

The terminal universal property of a functor l (as a corepresentable profunctor) and an object b.

Associated Types

type R (l :: j +-> k) (b :: j) :: k

Methods

termUnivArr :: l (R l b) b

termUnivProp :: forall (a :: k). l a b -> a ~> R l b

newtype AsRightAdjoint (r :: k -> k1 -> Type) (a :: k) (b :: k1)

Constructors

AsRightAdjoint

Fields

unAsRightAdjoint :: r a b

newtype AsLeftAdjoint (l :: k -> k1 -> Type) (a :: k) (b :: k1)

Constructors

AsLeftAdjoint

Fields

unAsLeftAdjoint :: l a b

newtype FromAdjunction (p :: k -> k1 -> Type) (a :: k) (b :: k1)

Constructors

FromAdjunction

Fields

unFromAdjunction :: p a b