Proarrow.Category.Bicategory

Bicategories

class BicategoryConstraints kk => Bicategory kk where

Bicategories.

0-cells are kinds i, j, k... satisfying the Ob0 kk constraint.

1-cells are types of kind kk j k for any 0-cells j and k, satisfying the Ob constraint.

2-cells are values of type p ~> q, where p and q are 1-cells.

Associated Types

type Ob0 kk (j :: k) :: Constraint

type Ob0 kk j = Any j

type I :: kk i i

type O (p :: kk i j) (q :: kk j k) :: kk i k

Methods

o :: ((a :: kk i j) ~> b) -> (c ~> d) -> (a `O` c) ~> (b `O` d)

Horizontal composition of 2-cells.

(\\\) :: ((Ob0 kk i, Ob0 kk j, Ob ps, Ob qs) => r) -> ((ps :: kk i j) ~> qs) -> r

Observe constraints from a 2-cell value.

leftUnitor :: Obj (a :: kk i j) -> (I `O` a) ~> a

leftUnitorInv :: Obj (a :: kk i j) -> a ~> (I `O` a)

rightUnitor :: Obj (a :: kk i j) -> (a `O` I) ~> a

rightUnitorInv :: Obj (a :: kk i j) -> a ~> (a `O` I)

associator :: Obj (a :: kk i j) -> Obj b -> Obj c -> ((a `O` b) `O` c) ~> (a `O` (b `O` c))

associatorInv :: Obj (a :: kk i j) -> Obj b -> Obj c -> (a `O` (b `O` c)) ~> ((a `O` b) `O` c)

class (forall j k. (Ob0 kk j, Ob0 kk k) => CategoryOf (kk j k), forall i. Ob0 kk i => IIsOb kk i) => BicategoryConstraints kk

class Ob (I :: kk i i) => IIsOb kk i

appendObj :: forall {kk} {i} {j} {k} (a :: kk i j) (b :: kk j k) r. (Bicategory kk, Ob0 kk i, Ob0 kk j, Ob0 kk k, Ob a, Ob b) => (Ob (a `O` b) => r) -> r

Paths

type data Path kk j k where

The type of 2-parameter kind constructors.

A type-level kind-threaded list. Used to strictify the bicategory and double category definitions. kk is a kind constructor, which creates a kind given two other kinds. (Each kind representing a 0-cell.)

Constructors

Nil :: Path kk k k

(:::) :: kk i j -> Path kk j k -> Path kk i k infixr 5

data SPath ps where

Constructors

SNil :: Ob0 kk k => SPath (Nil :: Path kk k k)

SCons :: forall {kk} {j} {k} (p :: kk j k) ps. (Ob0 kk j, Ob0 kk k) => Obj p -> SPath ps -> SPath (p ::: ps)

class IsPath (ps :: Path kk j k) where

Methods

singPath :: SPath ps

asObj :: Bicategory kk => SPath (ps :: Path kk i j) -> Obj ps

asSPath :: Bicategory kk => Obj (ps :: Path kk i j) -> SPath ps

type family (ps :: Path kk a b) +++ (qs :: Path kk b c) :: Path kk a c infixl 5

append :: SPath ps -> SPath qs -> SPath (ps +++ qs)

withAssoc :: forall {kk} {h} {i} {j} {k} (a :: Path kk h i) (b :: Path kk i j) (c :: Path kk j k) r. Bicategory kk => Obj a -> Obj b -> Obj c -> (((a +++ b) +++ c) ~ (a +++ (b +++ c)) => r) -> r

withUnital :: forall {kk} {j} {k} (a :: Path kk j k) r. Bicategory kk => Obj a -> ((a +++ Nil) ~ a => r) -> r

data Strictified ps qs where

Constructors

Str :: forall {kk} {j} {k} (ps :: Path kk j k) qs. (Ob ps, Ob qs, Ob0 kk j, Ob0 kk k) => (Fold ps ~> Fold qs) -> Strictified ps qs

type family Fold (as :: Path kk j k) :: kk j k

fold :: forall {kk} {i} {j} (as :: Path kk i j). Bicategory kk => SPath as -> Fold as ~> Fold as

obj1 :: forall {kk} {i} {j} p. (Bicategory kk, Ob (p :: kk i j), Ob0 kk i, Ob0 kk j) => Obj (p ::: Nil)

concatFold :: forall {kk} {i} {j} {k} (as :: Path kk i j) (bs :: Path kk j k). (Ob as, Ob bs, Ob0 kk i, Ob0 kk j, Ob0 kk k, Bicategory kk) => (Fold as `O` Fold bs) ~> Fold (as +++ bs)

splitFold :: forall {kk} {i} {j} {k} (as :: Path kk i j) (bs :: Path kk j k). (Ob as, Ob bs, Ob0 kk i, Ob0 kk j, Ob0 kk k, Bicategory kk) => Fold (as +++ bs) ~> (Fold as `O` Fold bs)

introI :: forall {kk} {j}. (Ob0 kk j, Bicategory kk) => (Nil :: Path kk j j) ~> (I ::: Nil)

elimI :: forall {kk} {j}. (Ob0 kk j, Bicategory kk) => (I ::: Nil) ~> (Nil :: Path kk j j)

introO :: forall {kk} {i} {j} {k} (p :: kk i j) (q :: kk j k). (Ob0 kk i, Ob0 kk j, Ob0 kk k, Bicategory kk, Ob p, Ob q) => (p ::: (q ::: Nil)) ~> ((p `O` q) ::: Nil)

elimO :: forall {kk} {i} {j} {k} (p :: kk i j) (q :: kk j k). (Ob0 kk i, Ob0 kk j, Ob0 kk k, Bicategory kk, Ob p, Ob q) => ((p `O` q) ::: Nil) ~> (p ::: (q ::: Nil))

More

class (Bicategory kk, Ob0 kk a, Ob t) => Monad (t :: kk a a) where

Methods

eta :: I ~> t

mu :: (t `O` t) ~> t

class (Bicategory kk, Ob0 kk a, Ob t) => Comonad (t :: kk a a) where

Methods

epsilon :: t ~> I

delta :: t ~> (t `O` t)

class (Bicategory kk, Ob0 kk c, Ob0 kk d, Ob l, Ob r) => Adjunction (l :: kk c d) (r :: kk d c) where

Methods

unit :: I ~> (r `O` l)

counit :: (l `O` r) ~> I

class (Monad s, Monad t) => Bimodule s t (p :: kk a b) where

Methods

leftAction :: (s `O` p) ~> p

rightAction :: (p `O` t) ~> p