Proarrow.Category.Bicategory

Bicategories

class BicategoryConstraints kk => Bicategory (kk :: CAT k) where

Bicategories.

0-cells are kinds i, j, k... satisfying the Ob0 kk constraint.

1-cells are types of kind kk j k for any 0-cells j and k, satisfying the Ob constraint.

2-cells are values of type p ~> q, where p and q are 1-cells.

Associated Types

type Ob0 (kk :: CAT k) (j :: k1)

type Ob0 (kk :: CAT k) (j :: k1) = Any j

type I :: kk i i

type O (p :: kk i j) (q :: kk j k1) :: kk i k1

Methods

o :: forall {k1 :: k} (i :: k) (j :: k) (a :: kk i j) (b :: kk i j) (c :: kk j k1) (d :: kk j k1). (a ~> b) -> (c ~> d) -> O a c ~> O b d

Horizontal composition of 2-cells.

(\\\) :: forall (i :: k) (j :: k) (ps :: kk i j) (qs :: kk i j) r. ((Ob0 kk i, Ob0 kk j, Ob ps, Ob qs) => r) -> (ps ~> qs) -> r

Observe constraints from a 2-cell value.

leftUnitor :: forall (i :: k) (j :: k) (a :: kk i j). Obj a -> O (I :: kk i i) a ~> a

leftUnitorInv :: forall (i :: k) (j :: k) (a :: kk i j). Obj a -> a ~> O (I :: kk i i) a

rightUnitor :: forall (i :: k) (j :: k) (a :: kk i j). Obj a -> O a (I :: kk j j) ~> a

rightUnitorInv :: forall (i :: k) (j :: k) (a :: kk i j). Obj a -> a ~> O a (I :: kk j j)

associator :: forall {j1 :: k} {k1 :: k} (i :: k) (j2 :: k) (a :: kk i j2) (b :: kk j2 j1) (c :: kk j1 k1). Obj a -> Obj b -> Obj c -> O (O a b) c ~> O a (O b c)

associatorInv :: forall {j1 :: k} {k1 :: k} (i :: k) (j2 :: k) (a :: kk i j2) (b :: kk j2 j1) (c :: kk j1 k1). Obj a -> Obj b -> Obj c -> O a (O b c) ~> O (O a b) c

class (forall (j :: k) (k1 :: k). (Ob0 kk j, Ob0 kk k1) => CategoryOf (kk j k1), forall (i :: k). Ob0 kk i => IIsOb kk i) => BicategoryConstraints (kk :: CAT k)

class Ob (I :: kk i i) => IIsOb (kk :: k -> k -> Type) (i :: k)

appendObj :: forall {k1} {kk :: k1 -> k1 -> Type} {i :: k1} {j :: k1} {k2 :: k1} (a :: kk i j) (b :: kk j k2) r. (Bicategory kk, Ob0 kk i, Ob0 kk j, Ob0 kk k2, Ob a, Ob b) => (Ob (O a b) => r) -> r

Paths

data Path (kk :: CAT k) (j :: k) (k1 :: k) where

The type of 2-parameter kind constructors.

A type-level kind-threaded list. Used to strictify the bicategory and double category definitions. kk is a kind constructor, which creates a kind given two other kinds. (Each kind representing a 0-cell.)

Constructors

Nil :: forall {k} (kk :: CAT k) (j :: k). Path kk j j

(:::) :: forall {k} (kk :: CAT k) (j :: k) (j1 :: k) (k1 :: k). kk j j1 -> Path kk j1 k1 -> Path kk j k1 infixr 5

data SPath (ps :: Path kk j k1) where

Constructors

SNil :: forall {k} (kk :: CAT k) (j :: k). Ob0 kk j => SPath ('Nil :: Path kk j j)

SCons :: forall {k} {k1 :: k} {kk :: CAT k} {j :: k} {k2 :: k} (p :: kk j k2) (ps1 :: Path kk k2 k1). (Ob0 kk j, Ob0 kk k2) => Obj p -> SPath ps1 -> SPath (p '::: ps1)

class IsPath (ps :: Path kk j k1) where

Methods

singPath :: SPath ps

asObj :: forall {k} (kk :: CAT k) (i :: k) (j :: k) (ps :: Path kk i j). Bicategory kk => SPath ps -> Obj ps

asSPath :: forall {k} (kk :: CAT k) (i :: k) (j :: k) (ps :: Path kk i j). Bicategory kk => Obj ps -> SPath ps

type family (ps :: Path kk a b) +++ (qs :: Path kk b c) :: Path kk a c infixl 5

append :: forall {k1} {kk :: CAT k1} {j :: k1} {b :: k1} {k2 :: k1} (ps :: Path kk j b) (qs :: Path kk b k2). SPath ps -> SPath qs -> SPath (ps +++ qs)

withAssoc :: forall {k1} {kk :: CAT k1} {h :: k1} {i :: k1} {j :: k1} {k2 :: k1} (a :: Path kk h i) (b :: Path kk i j) (c :: Path kk j k2) r. Bicategory kk => Obj a -> Obj b -> Obj c -> (((a +++ b) +++ c) ~ (a +++ (b +++ c)) => r) -> r

withUnital :: forall {k1} {kk :: CAT k1} {j :: k1} {k2 :: k1} (a :: Path kk j k2) r. Bicategory kk => Obj a -> ((a +++ ('Nil :: Path kk k2 k2)) ~ a => r) -> r

data Strictified (ps :: Path kk j k1) (qs :: Path kk j k1) where

Constructors

Str :: forall {k} {kk :: CAT k} {j :: k} {k1 :: k} (ps :: Path kk j k1) (qs :: Path kk j k1). (Ob ps, Ob qs, Ob0 kk j, Ob0 kk k1) => (Fold ps ~> Fold qs) -> Strictified ps qs

type family Fold (as :: Path kk j k1) :: kk j k1

fold :: forall {k} {kk :: CAT k} {i :: k} {j :: k} (as :: Path kk i j). Bicategory kk => SPath as -> Fold as ~> Fold as

obj1 :: forall {k} {kk :: CAT k} {i :: k} {j :: k} (p :: kk i j). (Bicategory kk, Ob p, Ob0 kk i, Ob0 kk j) => Obj (p '::: ('Nil :: Path kk j j))

concatFold :: forall {k1} {kk :: CAT k1} {i :: k1} {j :: k1} {k2 :: k1} (as :: Path kk i j) (bs :: Path kk j k2). (Ob as, Ob bs, Ob0 kk i, Ob0 kk j, Ob0 kk k2, Bicategory kk) => O (Fold as) (Fold bs) ~> Fold (as +++ bs)

splitFold :: forall {k1} {kk :: CAT k1} {i :: k1} {j :: k1} {k2 :: k1} (as :: Path kk i j) (bs :: Path kk j k2). (Ob as, Ob bs, Ob0 kk i, Ob0 kk j, Ob0 kk k2, Bicategory kk) => Fold (as +++ bs) ~> O (Fold as) (Fold bs)

introI :: forall {k} {kk :: CAT k} {j :: k}. (Ob0 kk j, Bicategory kk) => ('Nil :: Path kk j j) ~> ((I :: kk j j) '::: ('Nil :: Path kk j j))

elimI :: forall {k} {kk :: CAT k} {j :: k}. (Ob0 kk j, Bicategory kk) => ((I :: kk j j) '::: ('Nil :: Path kk j j)) ~> ('Nil :: Path kk j j)

introO :: forall {k1} {kk :: k1 -> k1 -> Type} {i :: k1} {j :: k1} {k2 :: k1} (p :: kk i j) (q :: kk j k2). (Ob0 kk i, Ob0 kk j, Ob0 kk k2, Bicategory kk, Ob p, Ob q) => (p '::: (q '::: ('Nil :: Path kk k2 k2))) ~> (O p q '::: ('Nil :: Path kk k2 k2))

elimO :: forall {k1} {kk :: k1 -> k1 -> Type} {i :: k1} {j :: k1} {k2 :: k1} (p :: kk i j) (q :: kk j k2). (Ob0 kk i, Ob0 kk j, Ob0 kk k2, Bicategory kk, Ob p, Ob q) => (O p q '::: ('Nil :: Path kk k2 k2)) ~> (p '::: (q '::: ('Nil :: Path kk k2 k2)))

More

class (Bicategory kk, Ob0 kk a, Ob t) => Monad (t :: kk a a) where

Methods

eta :: (I :: kk a a) ~> t

mu :: O t t ~> t

class (Bicategory kk, Ob0 kk a, Ob t) => Comonad (t :: kk a a) where

Methods

epsilon :: t ~> (I :: kk a a)

delta :: t ~> O t t

class (Bicategory kk, Ob0 kk c, Ob0 kk d, Ob l, Ob r) => Adjunction (l :: kk c d) (r :: kk d c) where

Methods

unit :: (I :: kk d d) ~> O r l

counit :: O l r ~> (I :: kk c c)

class (Monad s, Monad t) => Bimodule (s :: kk a a) (t :: kk b b) (p :: kk a b) where

Methods

leftAction :: O s p ~> p

rightAction :: O p t ~> p