Proarrow.Category.Bicategory

Bicategories

class (Locally CategoryOf kk, CategoryOf s, forall (i :: s). Ob0 kk i => ObUnit kk i) => Bicategory (kk :: CAT s) where

Bicategories.

0-cells are kinds i, j, k... satisfying the Ob0 kk constraint.

1-cells are types of kind kk j k for any 0-cells j and k, satisfying the Ob constraint.

2-cells are values of type p ~> q, where p and q are 1-cells.

Associated Types

type Ob0 (kk :: CAT s) (j :: k)

type Ob0 (kk :: CAT s) (j :: k) = ()

type I :: kk i i

type O (p :: kk j k) (q :: kk i j) :: kk i k

Methods

o :: forall {i :: s} (j :: s) (k :: s) (a :: kk j k) (b :: kk j k) (c :: kk i j) (d :: kk i j). (a ~> b) -> (c ~> d) -> O a c ~> O b d

Horizontal composition of 2-cells.

withOb2 :: forall {i :: s} {j :: s} {k :: s} (a :: kk j k) (b :: kk i j) r. (Ob a, Ob b, Ob0 kk i, Ob0 kk j, Ob0 kk k) => (Ob (O a b) => r) -> r

Get proof that the composition of 2 1-cells is also a 1-cell.

withOb0s :: forall {j :: s} {k :: s} (a :: kk j k) r. Ob a => ((Ob0 kk j, Ob0 kk k) => r) -> r

Get proof that the source and target of a 1-cell are 0-cells.

(\\\) :: forall (i :: s) (j :: s) (ps :: kk i j) (qs :: kk i j) r. ((Ob0 kk i, Ob0 kk j, Ob ps, Ob qs) => r) -> (ps ~> qs) -> r infixl 1

Observe constraints from a 2-cell value.

leftUnitor :: forall {i :: s} {j :: s} (a :: kk i j). (Ob0 kk i, Ob0 kk j, Ob a) => O (I :: kk j j) a ~> a

leftUnitorInv :: forall {i :: s} {j :: s} (a :: kk i j). (Ob0 kk i, Ob0 kk j, Ob a) => a ~> O (I :: kk j j) a

rightUnitor :: forall {i :: s} {j :: s} (a :: kk i j). (Ob0 kk i, Ob0 kk j, Ob a) => O a (I :: kk i i) ~> a

rightUnitorInv :: forall {i :: s} {j :: s} (a :: kk i j). (Ob0 kk i, Ob0 kk j, Ob a) => a ~> O a (I :: kk i i)

associator :: forall {h :: s} {i :: s} {j :: s} {k :: s} (a :: kk j k) (b :: kk i j) (c :: kk h i). (Ob0 kk h, Ob0 kk i, Ob0 kk j, Ob0 kk k, Ob a, Ob b, Ob c) => O (O a b) c ~> O a (O b c)

associatorInv :: forall {h :: s} {i :: s} {j :: s} {k :: s} (a :: kk j k) (b :: kk i j) (c :: kk h i). (Ob0 kk h, Ob0 kk i, Ob0 kk j, Ob0 kk k, Ob a, Ob b, Ob c) => O a (O b c) ~> O (O a b) c

iObj :: forall {s} (kk :: CAT s) (i :: s). (Bicategory kk, Ob0 kk i) => Obj (I :: kk i i)

The identity 1-cell (as represented by an identity 2-cell).

(==) :: forall k (a :: k) (b :: k) (c :: k). CategoryOf k => (a ~> b) -> (b ~> c) -> a ~> c infixl 7

(||) :: forall {s} {i1 :: s} (kk :: CAT s) (i2 :: s) (j :: s) (a :: kk i2 j) (b :: kk i2 j) (c :: kk i1 i2) (d :: kk i1 i2). Bicategory kk => (a ~> b) -> (c ~> d) -> O a c ~> O b d infixl 8

leftUnitor' :: forall {s} (kk :: CAT s) (i :: s) (j :: s) (a :: kk i j) (b :: kk i j). Bicategory kk => (a ~> b) -> O (I :: kk j j) a ~> b

leftUnitorInv' :: forall {s} (kk :: CAT s) (i :: s) (j :: s) (a :: kk i j) (b :: kk i j). Bicategory kk => (a ~> b) -> a ~> O (I :: kk j j) b

rightUnitor' :: forall {s} (kk :: CAT s) (i :: s) (j :: s) (a :: kk i j) (b :: kk i j). Bicategory kk => (a ~> b) -> O a (I :: kk i i) ~> b

rightUnitorInv' :: forall {s} (kk :: CAT s) (i :: s) (j :: s) (a :: kk i j) (b :: kk i j). Bicategory kk => (a ~> b) -> a ~> O b (I :: kk i i)

associator' :: forall {s} {j1 :: s} {i1 :: s} (kk :: s -> s -> Type) {i2 :: s} {j2 :: s} (a :: kk i2 j2) (b :: kk j1 i2) (c :: kk i1 j1). Bicategory kk => Obj a -> Obj b -> Obj c -> O (O a b) c ~> O a (O b c)

associatorInv' :: forall {s} {j1 :: s} {i1 :: s} (kk :: s -> s -> Type) {i2 :: s} {j2 :: s} (a :: kk i2 j2) (b :: kk j1 i2) (c :: kk i1 j1). Bicategory kk => Obj a -> Obj b -> Obj c -> O a (O b c) ~> O (O a b) c

leftUnitorWith :: forall {s} {i1 :: s} (kk :: CAT s) (i2 :: s) (c :: kk i2 i2) (a :: kk i1 i2) (b :: kk i1 i2). Bicategory kk => (c ~> (I :: kk i2 i2)) -> (a ~> b) -> O c a ~> b

leftUnitorInvWith :: forall {s} {i1 :: s} (kk :: CAT s) (i2 :: s) (c :: kk i2 i2) (a :: kk i1 i2) (b :: kk i1 i2). Bicategory kk => ((I :: kk i2 i2) ~> c) -> (a ~> b) -> a ~> O c b

rightUnitorWith :: forall {s} {k :: s} (kk :: CAT s) (i :: s) (c :: kk i i) (a :: kk i k) (b :: kk i k). Bicategory kk => (c ~> (I :: kk i i)) -> (a ~> b) -> O a c ~> b

rightUnitorInvWith :: forall {s} {k :: s} (kk :: CAT s) (i :: s) (c :: kk i i) (a :: kk i k) (b :: kk i k). Bicategory kk => ((I :: kk i i) ~> c) -> (a ~> b) -> a ~> O b c

class Ob0 kk j => Ob0' (kk :: CAT s) (j :: k)

class (Ob0 kk j, Ob0 kk k1, Ob a) => Ob' (a :: kk j k1)

More

class (Bicategory kk, Ob0 kk a, Ob t) => Monad (t :: kk a a) where

Methods

eta :: (I :: kk a a) ~> t

mu :: O t t ~> t

class (Bicategory kk, Ob0 kk a, Ob t) => Comonad (t :: kk a a) where

Methods

epsilon :: t ~> (I :: kk a a)

delta :: t ~> O t t

class (Bicategory kk, Ob0 kk c, Ob0 kk d) => Adjunction (l :: kk c d) (r :: kk d c) where

Methods

unit :: (I :: kk c c) ~> O r l

counit :: O l r ~> (I :: kk d d)

leftAdjunct :: forall {s} {kk :: s -> s -> Type} {c :: s} {d :: s} {i :: s} (l :: kk c d) (r :: kk d c) (a :: kk i c) (b :: kk i d). (Adjunction l r, Ob a, Ob r, Ob l, Ob0 kk i) => (O l a ~> b) -> a ~> O r b

rightAdjunct :: forall {s} {kk :: s -> s -> Type} {c :: s} {d :: s} {i :: s} (l :: kk c d) (r :: kk d c) (a :: kk i c) (b :: kk i d). (Adjunction l r, Ob b, Ob r, Ob l, Ob0 kk i) => (a ~> O r b) -> O l a ~> b

class (Monad s, Monad t) => Bimodule (s :: kk a a) (t :: kk b b) (p :: kk b a) where

Methods

leftAction :: O s p ~> p

rightAction :: O p t ~> p