Proarrow.Category.Bicategory

Bicategories

class (Locally CategoryOf kk, CategoryOf s) => Bicategory (kk :: CAT s) where

Bicategories.

0-cells are kinds i, j, k... satisfying the Ob0 kk constraint.

1-cells are types of kind kk j k for any 0-cells j and k, satisfying the Ob constraint.

2-cells are values of type p ~> q, where p and q are 1-cells.

Minimal complete definition

o, (\\\), leftUnitor, leftUnitorInv, rightUnitor, rightUnitorInv, associator, associatorInv

Associated Types

type Ob0 (kk :: CAT s) (j :: k)

type Ob0 (kk :: CAT s) (j :: k) = ()

type I :: kk i i

type O (p :: kk j k) (q :: kk i j) :: kk i k

Methods

iObj :: forall (i :: s). Ob0 kk i => Obj (I :: kk i i)

The identity 1-cell (as represented by an identity 2-cell).

default iObj :: forall (i :: s). (Ob0 kk i, Ob (I :: kk i i)) => Obj (I :: kk i i)

o :: forall {i :: s} (j :: s) (k :: s) (a :: kk j k) (b :: kk j k) (c :: kk i j) (d :: kk i j). (a ~> b) -> (c ~> d) -> O a c ~> O b d

Horizontal composition of 2-cells.

(\\\) :: forall (i :: s) (j :: s) (ps :: kk i j) (qs :: kk i j) r. ((Ob0 kk i, Ob0 kk j, Ob ps, Ob qs) => r) -> (ps ~> qs) -> r

Observe constraints from a 2-cell value.

leftUnitor :: forall (i :: s) (j :: s) (a :: kk i j) (b :: kk i j). (a ~> b) -> O (I :: kk j j) a ~> b

leftUnitorInv :: forall (i :: s) (j :: s) (a :: kk i j) (b :: kk i j). (a ~> b) -> a ~> O (I :: kk j j) b

rightUnitor :: forall (i :: s) (j :: s) (a :: kk i j) (b :: kk i j). (a ~> b) -> O a (I :: kk i i) ~> b

rightUnitorInv :: forall (i :: s) (j :: s) (a :: kk i j) (b :: kk i j). (a ~> b) -> a ~> O b (I :: kk i i)

associator :: forall {j1 :: s} {i1 :: s} (i2 :: s) (j2 :: s) (a :: kk i2 j2) (b :: kk j1 i2) (c :: kk i1 j1). Obj a -> Obj b -> Obj c -> O (O a b) c ~> O a (O b c)

associatorInv :: forall {j1 :: s} {i1 :: s} (i2 :: s) (j2 :: s) (a :: kk i2 j2) (b :: kk j1 i2) (c :: kk i1 j1). Obj a -> Obj b -> Obj c -> O a (O b c) ~> O (O a b) c

(==) :: forall k (a :: k) (b :: k) (c :: k). CategoryOf k => (a ~> b) -> (b ~> c) -> a ~> c infixl 7

(||) :: forall {s} {i1 :: s} (kk :: CAT s) (i2 :: s) (j :: s) (a :: kk i2 j) (b :: kk i2 j) (c :: kk i1 i2) (d :: kk i1 i2). Bicategory kk => (a ~> b) -> (c ~> d) -> O a c ~> O b d infixl 8

appendObj :: forall {s} {kk :: s -> s -> Type} {i :: s} {j :: s} {k :: s} (a :: kk j k) (b :: kk i j) r. (Bicategory kk, Ob0 kk i, Ob0 kk j, Ob0 kk k, Ob a, Ob b) => (Ob (O a b) => r) -> r

leftUnitorWith :: forall {s} {i1 :: s} (kk :: CAT s) (i2 :: s) (c :: kk i2 i2) (a :: kk i1 i2) (b :: kk i1 i2). Bicategory kk => (c ~> (I :: kk i2 i2)) -> (a ~> b) -> O c a ~> b

leftUnitorInvWith :: forall {s} {i1 :: s} (kk :: CAT s) (i2 :: s) (c :: kk i2 i2) (a :: kk i1 i2) (b :: kk i1 i2). Bicategory kk => ((I :: kk i2 i2) ~> c) -> (a ~> b) -> a ~> O c b

rightUnitorWith :: forall {s} {k :: s} (kk :: CAT s) (i :: s) (c :: kk i i) (a :: kk i k) (b :: kk i k). Bicategory kk => (c ~> (I :: kk i i)) -> (a ~> b) -> O a c ~> b

rightUnitorInvWith :: forall {s} {k :: s} (kk :: CAT s) (i :: s) (c :: kk i i) (a :: kk i k) (b :: kk i k). Bicategory kk => ((I :: kk i i) ~> c) -> (a ~> b) -> a ~> O b c

Paths

data Path (kk :: CAT k) (j :: k) (k1 :: k) where

The type of 2-parameter kind constructors.

A type-level kind-threaded list. Used to strictify the bicategory and double category definitions. kk is a kind constructor, which creates a kind given two other kinds. (Each kind representing a 0-cell.)

Constructors

Nil :: forall {k} (kk :: CAT k) (j :: k). Path kk j j

(:::) :: forall {k} (kk :: CAT k) (j :: k) (j1 :: k) (k1 :: k). kk j j1 -> Path kk j1 k1 -> Path kk j k1 infixr 5

data SPath (ps :: Path kk j k1) where

Constructors

SNil :: forall {k} (kk :: CAT k) (j :: k). Ob0 kk j => SPath ('Nil :: Path kk j j)

SCons :: forall {k} {kk :: CAT k} {j :: k} {j1 :: k} {k1 :: k} (p :: kk j j1) (ps1 :: Path kk j1 k1). Ob0 kk j => Obj p -> SPath ps1 -> SPath (p '::: ps1)

class (Bicategory kk, Ob0 kk j, Ob0 kk k1) => IsPath (ps :: Path kk j k1) where

Methods

singPath :: SPath ps

withIsPath :: forall {k1} (kk :: CAT k1) (j :: k1) (k2 :: k1) (ps :: Path kk j k2) r. Bicategory kk => SPath ps -> ((IsPath ps, Ob0 kk j, Ob0 kk k2) => r) -> r

asObj :: forall {k} (kk :: CAT k) (i :: k) (j :: k) (ps :: Path kk i j). Bicategory kk => SPath ps -> Obj ps

asSPath :: forall {k} (kk :: CAT k) (i :: k) (j :: k) (ps :: Path kk i j). Bicategory kk => Obj ps -> SPath ps

type family (ps :: Path kk a b) +++ (qs :: Path kk b c) :: Path kk a c infixl 5

append :: forall {k1} {kk :: CAT k1} {j :: k1} {b :: k1} {k2 :: k1} (ps :: Path kk j b) (qs :: Path kk b k2). SPath ps -> SPath qs -> SPath (ps +++ qs)

singleton :: forall {k} (kk :: CAT k) (i :: k) (j :: k) (p :: kk i j) (q :: kk i j). Bicategory kk => (p ~> q) -> (p '::: ('Nil :: Path kk j j)) ~> (q '::: ('Nil :: Path kk j j))

withAssoc :: forall {k1} {kk :: CAT k1} {h :: k1} {i :: k1} {j :: k1} {k2 :: k1} (a :: Path kk h i) (b :: Path kk i j) (c :: Path kk j k2) r. Bicategory kk => Obj a -> Obj b -> Obj c -> (((a +++ b) +++ c) ~ (a +++ (b +++ c)) => r) -> r

withUnital :: forall {k1} {kk :: CAT k1} {j :: k1} {k2 :: k1} (a :: Path kk j k2) r. Bicategory kk => SPath a -> ((a +++ ('Nil :: Path kk k2 k2)) ~ a => r) -> r

data Strictified (ps :: Path kk j k1) (qs :: Path kk j k1) where

Constructors

Str :: forall {k} {kk :: CAT k} {j :: k} {k1 :: k} (ps :: Path kk j k1) (qs :: Path kk j k1). (Ob0 kk j, Ob0 kk k1) => SPath ps -> SPath qs -> (Fold ps ~> Fold qs) -> Strictified ps qs

str :: forall {k1} (kk :: CAT k1) (j :: k1) (k2 :: k1) (ps :: Path kk j k2) (qs :: Path kk j k2). Bicategory kk => Obj ps -> Obj qs -> (Fold ps ~> Fold qs) -> Strictified ps qs

unStr :: forall {k1} {kk :: CAT k1} {j :: k1} {k2 :: k1} (ps :: Path kk j k2) (qs :: Path kk j k2). Strictified ps qs -> Fold ps ~> Fold qs

type family Fold (as :: Path kk j k1) :: kk j k1

fold :: forall {k} {kk :: CAT k} {i :: k} {j :: k} (as :: Path kk i j). Bicategory kk => SPath as -> Fold as ~> Fold as

obj1 :: forall {k} {kk :: CAT k} {i :: k} {j :: k} (p :: kk i j). (Bicategory kk, Ob p, Ob0 kk i, Ob0 kk j) => Obj (p '::: ('Nil :: Path kk j j))

concatFold :: forall {k1} {kk :: CAT k1} {i :: k1} {j :: k1} {k2 :: k1} (as :: Path kk i j) (bs :: Path kk j k2). Bicategory kk => SPath as -> SPath bs -> O (Fold bs) (Fold as) ~> Fold (as +++ bs)

splitFold :: forall {k1} {kk :: CAT k1} {i :: k1} {j :: k1} {k2 :: k1} (as :: Path kk i j) (bs :: Path kk j k2). Bicategory kk => SPath as -> SPath bs -> Fold (as +++ bs) ~> O (Fold bs) (Fold as)

introI :: forall {k} {kk :: CAT k} {j :: k}. (Ob0 kk j, Bicategory kk) => ('Nil :: Path kk j j) ~> ((I :: kk j j) '::: ('Nil :: Path kk j j))

elimI :: forall {s} {kk :: CAT s} {j :: s}. (Ob0 kk j, Bicategory kk) => ((I :: kk j j) '::: ('Nil :: Path kk j j)) ~> ('Nil :: Path kk j j)

introO :: forall {k1} {kk :: k1 -> k1 -> Type} {i :: k1} {j :: k1} {k2 :: k1} (p :: kk i j) (q :: kk j k2). (Ob0 kk i, Ob0 kk j, Ob0 kk k2, Bicategory kk, Ob p, Ob q) => (p '::: (q '::: ('Nil :: Path kk k2 k2))) ~> (O q p '::: ('Nil :: Path kk k2 k2))

elimO :: forall {k1} {kk :: k1 -> k1 -> Type} {i :: k1} {j :: k1} {k2 :: k1} (p :: kk i j) (q :: kk j k2). (Ob0 kk i, Ob0 kk j, Ob0 kk k2, Bicategory kk, Ob p, Ob q) => (O q p '::: ('Nil :: Path kk k2 k2)) ~> (p '::: (q '::: ('Nil :: Path kk k2 k2)))

More

class (Bicategory kk, Ob0 kk a, Ob t) => Monad (t :: kk a a) where

Methods

eta :: (I :: kk a a) ~> t

mu :: O t t ~> t

class (Bicategory kk, Ob0 kk a, Ob t) => Comonad (t :: kk a a) where

Methods

epsilon :: t ~> (I :: kk a a)

delta :: t ~> O t t

class (Bicategory kk, Ob0 kk c, Ob0 kk d, Ob l, Ob r) => Adjunction (l :: kk c d) (r :: kk d c) where

Methods

unit :: (I :: kk c c) ~> O r l

counit :: O l r ~> (I :: kk d d)

leftAdjunct :: forall {k1} {kk :: CAT k1} {k2 :: k1} {j :: k1} {i :: k1} (l :: kk k2 j) (r :: kk j k2) (a :: kk i k2) (b :: kk i j). (Adjunction l r, Ob a) => (O l a ~> b) -> a ~> O r b

rightAdjunct :: forall {k1} {kk :: CAT k1} {j :: k1} {k2 :: k1} {i :: k1} (l :: kk j k2) (r :: kk k2 j) (a :: kk i j) (b :: kk i k2). (Adjunction l r, Ob b) => (a ~> O r b) -> O l a ~> b

class (Monad s, Monad t, Ob p) => Bimodule (s :: kk a a) (t :: kk b b) (p :: kk b a) where

Methods

leftAction :: O s p ~> p

rightAction :: O p t ~> p