Proarrow.Category.Double

Documentation

type SQ1 (hk :: CAT c) (vk :: CAT c) = forall {h :: c} {i :: c} {j :: c} {k :: c}. hk h j -> hk i k -> vk h i -> vk j k -> Type

The kind of a square p q f g. > h--f--i > | v | > p--@--q > | v | > j--g--k

type SQ (hk :: CAT c) (vk :: CAT c) = SQ1 (Path hk) (Path vk)

class (Bicategory hk, Bicategory vk) => Double (hk :: CAT k) (vk :: CAT k) where

Double categories, strictified in both directions.

Associated Types

data Sq (hk :: CAT k) (vk :: CAT k) :: SQ hk vk

data Sq1 (hk :: CAT k) (vk :: CAT k) :: SQ1 hk vk

Methods

singleton :: forall (h :: k) (i :: k) (j :: k) (k1 :: k) (p :: hk h j) (q :: hk i k1) (f :: vk h i) (g :: vk j k1). (DOb hk vk h, DOb hk vk i, DOb hk vk j, DOb hk vk k1, Ob p, Ob q, Ob f, Ob g) => Sq1 hk vk p q f g -> Sq hk vk (p '::: ('Nil :: Path hk j j)) (q '::: ('Nil :: Path hk k1 k1)) (f '::: ('Nil :: Path vk i i)) (g '::: ('Nil :: Path vk k1 k1))

folded :: forall (h :: k) (i :: k) (j :: k) (k1 :: k) (ps :: Path hk h j) (qs :: Path hk i k1) (fs :: Path vk h i) (gs :: Path vk j k1). (DOb hk vk h, DOb hk vk i, DOb hk vk j, DOb hk vk k1, Ob ps, Ob qs, Ob fs, Ob gs) => Sq hk vk ps qs fs gs -> Sq1 hk vk (Fold ps) (Fold qs) (Fold fs) (Fold gs)

object :: forall (k1 :: k). DOb hk vk k1 => Sq hk vk ('Nil :: Path hk k1 k1) ('Nil :: Path hk k1 k1) ('Nil :: Path vk k1 k1) ('Nil :: Path vk k1 k1)

The empty square for an object.

hArr :: forall (j :: k) (k1 :: k) (ps :: Path hk j k1) (qs :: Path hk j k1). (DOb hk vk j, DOb hk vk k1) => (ps ~> qs) -> Sq hk vk ps qs ('Nil :: Path vk j j) ('Nil :: Path vk k1 k1)

Make a square from a horizontal arrow

hArr1 :: forall (j :: k) (k1 :: k) (p :: hk j k1) (q :: hk j k1). (DOb hk vk j, DOb hk vk k1) => (p ~> q) -> Sq1 hk vk p q (I :: vk j j) (I :: vk k1 k1)

(|||) :: forall {a1 :: k} {a2 :: k} {b1 :: k} {b2 :: k} {c1 :: k} {c2 :: k} (ps :: Path hk a1 a2) (qs :: Path hk b1 b2) (fs :: Path vk a1 b1) (gs :: Path vk a2 b2) (rs :: Path hk c1 c2) (hs :: Path vk b1 c1) (is :: Path vk b2 c2). Sq hk vk ps qs fs gs -> Sq hk vk qs rs hs is -> Sq hk vk ps rs (fs +++ hs) (gs +++ is) infixl 6

Horizontal composition

vArr :: forall (j :: k) (k1 :: k) (fs :: Path vk j k1) (gs :: Path vk j k1). (DOb hk vk j, DOb hk vk k1) => (fs ~> gs) -> Sq hk vk ('Nil :: Path hk j j) ('Nil :: Path hk k1 k1) fs gs

Make a square from a vertical arrow (in the opposite direction to match the quintet construction)

vArr1 :: forall (j :: k) (k1 :: k) (f :: vk j k1) (g :: vk j k1). (DOb hk vk j, DOb hk vk k1) => (f ~> g) -> Sq1 hk vk (I :: hk j j) (I :: hk k1 k1) f g

(===) :: forall {h :: k} {b1 :: k} {i :: k} {b2 :: k} {j :: k} {k1 :: k} (ps :: Path hk h b1) (qs :: Path hk i b2) (fs :: Path vk h i) (gs :: Path vk b1 b2) (rs :: Path hk b1 j) (ss :: Path hk b2 k1) (hs :: Path vk j k1). Sq hk vk ps qs fs gs -> Sq hk vk rs ss gs hs -> Sq hk vk (ps +++ rs) (qs +++ ss) fs hs infixl 5

Vertical composition

(\\\\) :: forall (h :: k) (i :: k) (j :: k) (k1 :: k) (ps :: Path hk h j) (qs :: Path hk i k1) (fs :: Path vk h i) (gs :: Path vk j k1) r. ((DOb hk vk h, DOb hk vk i, DOb hk vk j, DOb hk vk k1, Ob ps, Ob qs, Ob fs, Ob gs) => r) -> Sq hk vk ps qs fs gs -> r

type DOb (hk :: CAT k) (vk :: CAT k1) (k3 :: k2) = (Ob0 hk k3, Ob0 vk k3)

class Double hk vk => Equipment (hk :: k -> k -> Type) (vk :: k -> k -> Type) where

Associated Types

type Companion (hk :: k -> k -> Type) (vk :: k -> k -> Type) (f :: vk j k1) :: hk j k1

type Conjoint (hk :: k -> k -> Type) (vk :: k -> k -> Type) (f :: vk j k1) :: hk k1 j

Methods

fromRight :: forall (j :: k) (k1 :: k) (f :: vk j k1). (DOb hk vk j, DOb hk vk k1, Ob f) => Sq hk vk ('Nil :: Path hk j j) (Companion hk vk f '::: ('Nil :: Path hk k1 k1)) ('Nil :: Path vk j j) (f '::: ('Nil :: Path vk k1 k1))

toLeft :: forall (j :: k) (k1 :: k) (f :: vk j k1). (DOb hk vk j, DOb hk vk k1, Ob f) => Sq hk vk (Companion hk vk f '::: ('Nil :: Path hk k1 k1)) ('Nil :: Path hk k1 k1) (f '::: ('Nil :: Path vk k1 k1)) ('Nil :: Path vk k1 k1)

toRight :: forall (j :: k) (k1 :: k) (f :: vk j k1). (DOb hk vk j, DOb hk vk k1, Ob f) => Sq hk vk ('Nil :: Path hk j j) (Conjoint hk vk f '::: ('Nil :: Path hk j j)) (f '::: ('Nil :: Path vk k1 k1)) ('Nil :: Path vk j j)

fromLeft :: forall (j :: k) (k1 :: k) (f :: vk j k1). (DOb hk vk j, DOb hk vk k1, Ob f) => Sq hk vk (Conjoint hk vk f '::: ('Nil :: Path hk j j)) ('Nil :: Path hk k1 k1) ('Nil :: Path vk k1 k1) (f '::: ('Nil :: Path vk k1 k1))

withComOb :: forall (j :: k) (k1 :: k) (f :: vk j k1) r. (DOb hk vk j, DOb hk vk k1, Ob f) => Obj f -> (Ob (Companion hk vk f) => r) -> r

withConOb :: forall (j :: k) (k1 :: k) (f :: vk j k1) r. (DOb hk vk j, DOb hk vk k1, Ob f) => Obj f -> (Ob (Conjoint hk vk f) => r) -> r

type UnOp (ps :: Path (OPK kk) j k1) = UN ('OP :: kk k1 j -> OPK kk j k1) (Fold ps)

type UnCo (fs :: Path (COK kk) j k1) = UN ('CO :: kk j k1 -> COK kk j k1) (Fold fs)

type CoSq (ps :: Path (OPK hk) h j) (qs :: Path (OPK hk) i k1) (fs :: Path (COK vk) h i) (gs :: Path (COK vk) j k1) = Sq (OPK hk) (COK vk) ps qs fs gs