Proarrow.Category.Equipment

Documentation

class Ob0 kk k1 => Ob0' (kk :: CAT s) (k1 :: k)

class (Bicategory hk, Bicategory vk, forall (k :: c). Ob0 vk k => Ob0' hk k) => HasCompanions (hk :: CAT c) (vk :: CAT c) | hk -> vk where

A double category with companions.

Associated Types

type Companion (hk :: CAT c) (vk :: CAT c) (f :: vk j k) :: hk j k

Methods

mapCompanion :: forall {j :: c} {k :: c} (f :: vk j k) (g :: vk j k). (f ~> g) -> Companion hk vk g ~> Companion hk vk f

compToId :: forall (k :: c). Ob0 vk k => Companion hk vk (I :: vk k k) ~> (I :: hk k k)

compFromId :: forall (k :: c). Ob0 vk k => (I :: hk k k) ~> Companion hk vk (I :: vk k k)

compToCompose :: forall {k1 :: c} {k2 :: c} {i :: c} (f :: vk k1 k2) (g :: vk i k1). Obj f -> Obj g -> Companion hk vk (O f g) ~> O (Companion hk vk f) (Companion hk vk g)

compFromCompose :: forall {j1 :: c} {k :: c} {j2 :: c} (f :: vk j1 k) (g :: vk j2 j1). Obj f -> Obj g -> O (Companion hk vk f) (Companion hk vk g) ~> Companion hk vk (O f g)

type SQ' (hk :: CAT c) (vk :: CAT c) (h :: c) (i :: c) (j :: c) (k :: c) = PRO (hk h j, vk h i) (hk i k, vk j k)

The kind of a square '(p, f) '(q, g).

h--f--i
|  v  |
p--@--q
|  v  |
j--g--k

type SQ (hk :: CAT c) (vk :: CAT c) = forall {h :: c} {i :: c} {j :: c} {k :: c}. SQ' hk vk h i j k

data Sq (pf :: (hk h j, vk h i)) (qg :: (hk i k, vk j k)) where

Constructors

Sq :: forall {c} {hk :: CAT c} {vk :: CAT c} {h :: c} {i :: c} {j :: c} {k :: c} (p :: hk h j) (q :: hk i k) (f :: vk h i) (g :: vk j k). (Ob0 vk h, Ob0 vk i, Ob0 vk j, Ob0 vk k, Ob p, Ob q, Ob f, Ob g) => (O (Companion hk vk g) p ~> O q (Companion hk vk f)) -> Sq '(p, f) '(q, g)

object :: forall {s} hk vk (k :: s). (HasCompanions hk vk, Ob0 vk k) => Sq '(I :: hk k k, I :: vk k k) '(I :: hk k k, I :: vk k k)

The empty square for an object.

hArr :: forall {s} {hk} {vk} {j :: s} {k :: s} (p :: hk j k) (q :: hk j k). (HasCompanions hk vk, Ob0 vk j, Ob0 vk k) => (p ~> q) -> Sq '(p, I :: vk j j) '(q, I :: vk k k)

Make a square from a horizontal arrow

(|||) :: forall {c} {k1 :: c} {hk} {vk} {h :: c} {i :: c} {j :: c} {k2 :: c} {l :: c} (p :: hk h j) (q :: hk i k2) (r :: hk l k1) (f :: vk h i) (g :: vk j k2) (d :: vk i l) (e :: vk k2 k1). HasCompanions hk vk => Sq '(p, f) '(q, g) -> Sq '(q, d) '(r, e) -> Sq '(p, O d f) '(r, O e g) infixl 6

Horizontal composition

vArr :: forall {s} {hk} {vk} {j :: s} {k :: s} (f :: vk j k) (g :: vk j k). (HasCompanions hk vk, Ob0 vk j, Ob0 vk k) => (f ~> g) -> Sq '(I :: hk j j, f) '(I :: hk k k, g)

Make a square from a vertical arrow

(===) :: forall {c} {k1 :: c} {hk} {vk} {h :: c} {i :: c} {j :: c} {k2 :: c} {l :: c} (p :: hk h j) (q :: hk i l) (r :: hk j k2) (s :: hk l k1) (f :: vk h i) (g :: vk j l) (e :: vk k2 k1). HasCompanions hk vk => Sq '(p, f) '(q, g) -> Sq '(r, g) '(s, e) -> Sq '(O r p, f) '(O s q, e) infixl 5

Vertical composition

fromRight :: forall {c} {hk} {vk} {j :: c} {k :: c} (f :: vk j k). (HasCompanions hk vk, Ob0 vk j, Ob0 vk k, Ob f) => Sq '(I :: hk j j, I :: vk j j) '(Companion hk vk f, f)

toLeft :: forall {s} {hk} {vk} {j :: s} {k :: s} (f :: vk j k). (HasCompanions hk vk, Ob0 vk j, Ob0 vk k, Ob f) => Sq '(Companion hk vk f, f) '(I :: hk k k, I :: vk k k)

toRight :: forall {s} {hk} {vk} {j :: s} {k :: s} (f :: vk j k). (Equipment hk vk, Ob0 vk j, Ob0 vk k, Ob f) => Sq '(I :: hk j j, f) '(Conjoint hk vk f, I :: vk j j)

fromLeft :: forall {s} {hk} {vk} {j :: s} {k :: s} (f :: vk j k). (Equipment hk vk, Ob0 vk j, Ob0 vk k, Ob f) => Sq '(Conjoint hk vk f, I :: vk k k) '(I :: hk k k, f)

flipCompanion :: forall {c} {k :: c} {j :: c} {i :: c} (hk :: CAT c) (vk :: CAT c) (f :: vk k j) (p :: hk i k) (q :: hk i j). (Equipment hk vk, Ob p) => Obj f -> (O (Companion hk vk f) p ~> q) -> p ~> O (Conjoint hk vk f) q

flipCompanionInv :: forall {c} {j :: c} {k :: c} {i :: c} (hk :: CAT c) (vk :: CAT c) (f :: vk j k) (p :: hk i j) (q :: hk i k). (Equipment hk vk, Ob q) => Obj f -> (p ~> O (Conjoint hk vk f) q) -> O (Companion hk vk f) p ~> q

flipConjoint :: forall {c} {i :: c} {k1 :: c} {k2 :: c} (hk :: CAT c) (vk :: CAT c) (f :: vk i k1) (p :: hk i k2) (q :: hk k1 k2). (Equipment hk vk, Ob p) => Obj f -> (O p (Conjoint hk vk f) ~> q) -> p ~> O q (Companion hk vk f)

flipConjointInv :: forall {c} {j :: c} {i :: c} {k :: c} (hk :: CAT c) (vk :: CAT c) (f :: vk j i) (p :: hk j k) (q :: hk i k). (Equipment hk vk, Ob q) => Obj f -> (p ~> O q (Companion hk vk f)) -> O p (Conjoint hk vk f) ~> q

class Adjunction (Companion hk vk f) (Conjoint hk vk f) => ComConAdjunction (hk :: CAT c) (vk :: CAT c) (f :: vk j k)

class HasCompanions hk vk => Equipment (hk :: CAT c) (vk :: CAT c) where

Minimal complete definition

mapConjoint

Associated Types

type Conjoint (hk :: CAT c) (vk :: CAT c) (f :: vk j k) :: hk k j

Methods

mapConjoint :: forall {j :: c} {k :: c} (f :: vk j k) (g :: vk j k). (f ~> g) -> Conjoint hk vk f ~> Conjoint hk vk g

conjToId :: forall (k :: c). Ob0 vk k => Conjoint hk vk (I :: vk k k) ~> (I :: hk k k)

conjFromId :: forall (k :: c). Ob0 vk k => (I :: hk k k) ~> Conjoint hk vk (I :: vk k k)

conjToCompose :: forall {j :: c} {i :: c} {k :: c} (f :: vk j i) (g :: vk k j). Obj f -> Obj g -> Conjoint hk vk (O f g) ~> O (Conjoint hk vk g) (Conjoint hk vk f)

conjFromCompose :: forall {j1 :: c} {k :: c} {j2 :: c} (f :: vk j1 k) (g :: vk j2 j1). Obj f -> Obj g -> O (Conjoint hk vk g) (Conjoint hk vk f) ~> Conjoint hk vk (O f g)

comConUnit :: forall {i :: c} {k :: c} (f :: vk i k). Obj f -> (I :: hk i i) ~> O (Conjoint hk vk f) (Companion hk vk f)

default comConUnit :: forall {i :: c} {k :: c} (f :: vk i k). (Ob f => ComConAdjunction hk vk f) => Obj f -> (I :: hk i i) ~> O (Conjoint hk vk f) (Companion hk vk f)

comConCounit :: forall {j :: c} {i :: c} (f :: vk j i). Obj f -> O (Companion hk vk f) (Conjoint hk vk f) ~> (I :: hk i i)

default comConCounit :: forall {j :: c} {i :: c} (f :: vk j i). (Ob f => ComConAdjunction hk vk f) => Obj f -> O (Companion hk vk f) (Conjoint hk vk f) ~> (I :: hk i i)

type Cart (p :: hk b d) (f :: vk a b) (g :: vk c1 d) = O (O (Conjoint hk vk g) p) (Companion hk vk f)

P(f, g)

companionFold :: forall {k1} {hk :: CAT k1} {vk :: CAT k1} {j :: k1} {k2 :: k1} (fs :: Path vk j k2). HasCompanions hk vk => SPath fs -> Companion hk vk (Fold fs) ~> Fold (Companion (Path hk) (Path vk) fs)

foldCompanion :: forall {k1} {hk :: CAT k1} {vk :: CAT k1} {j :: k1} {k2 :: k1} (fs :: Path vk j k2). HasCompanions hk vk => SPath fs -> Fold (Companion (Path hk) (Path vk) fs) ~> Companion hk vk (Fold fs)

mapCompanionSPath :: forall {k1} (hk :: CAT k1) {vk :: CAT k1} {j :: k1} {k2 :: k1} (fs :: Path vk j k2). HasCompanions hk vk => SPath fs -> SPath (Companion (Path hk) (Path vk) fs)

mapConjointSPath :: forall {k1} (hk :: CAT k1) {vk :: CAT k1} {j :: k1} {k2 :: k1} (fs :: Path vk j k2). Equipment hk vk => SPath fs -> SPath (Conjoint (Path hk) (Path vk) fs)

data RetroSq (pf :: (hk i k, vk j k)) (qg :: (hk h j, vk h i)) where

The kind of a retro square '(q, g) '(p, f).

h--f--i
|  v  |
p----q
|  v  |
j--g--k

Constructors

RetroSq :: forall {c} {hk :: CAT c} {vk :: CAT c} {h :: c} {i :: c} {j :: c} {k :: c} (p :: hk h j) (q :: hk i k) (f :: vk h i) (g :: vk j k). (Ob0 vk h, Ob0 vk i, Ob0 vk j, Ob0 vk k, Ob p, Ob q, Ob f, Ob g) => (O q (Companion hk vk f) ~> O (Companion hk vk g) p) -> RetroSq '(q, g) '(p, f)

adjVK :: forall {k1} {j1 :: k1} hk vk (i :: k1) (j2 :: k1) (k2 :: k1) (f :: vk j2 k2) (g :: vk j1 i) (v :: hk j2 j1) (w :: hk k2 i) (x :: hk j1 j2) (y :: hk i k2). (Adjunction x v, Adjunction y w, HasCompanions hk vk) => RetroSq '(y, f) '(x, g) -> Sq '(v, f) '(w, g)

adjHK :: forall {c} {i1 :: c} hk vk (i2 :: c) (j :: c) (k :: c) (e :: vk j k) (f :: vk i2 i1) (g :: vk k j) (h :: vk i1 i2) (v :: hk i2 k) (w :: hk i1 j). (Adjunction h f, Adjunction e g, HasCompanions hk vk) => RetroSq '(v, e) '(w, h) -> Sq '(v, f) '(w, g)

adj4Sq :: forall {c} {h1 :: c} hk vk (i :: c) (j :: c) (k :: c) (e :: vk h1 j) (f :: vk k i) (g :: vk j h1) (h2 :: vk i k) (v :: hk k j) (w :: hk i h1) (x :: hk j k) (y :: hk h1 i). (Adjunction v x, Adjunction w y, Adjunction h2 f, Adjunction e g, HasCompanions hk vk) => Sq '(v, f) '(w, g) -> Sq '(y, e) '(x, h2)