Proarrow.Category.Equipment

Documentation

class (Bicategory hk, Bicategory vk, forall (k :: c). Ob0 vk k => Ob0' hk k) => HasCompanions (hk :: CAT c) (vk :: CAT c) | hk -> vk where

A double category with companions.

Associated Types

type Companion (hk :: CAT c) (f :: vk j k) :: hk j k

Methods

mapCompanion :: forall {j :: c} {k :: c} (f :: vk j k) (g :: vk j k). (f ~> g) -> Companion hk f ~> Companion hk g

withObCompanion :: forall {j :: c} {k :: c} (f :: vk j k) r. Ob f => (Ob (Companion hk f) => r) -> r

compToId :: forall (k :: c). Ob0 vk k => Companion hk (I :: vk k k) ~> (I :: hk k k)

compFromId :: forall (k :: c). Ob0 vk k => (I :: hk k k) ~> Companion hk (I :: vk k k)

compToCompose :: forall {i :: c} {j :: c} {k :: c} (f :: vk j k) (g :: vk i j). Obj f -> Obj g -> Companion hk (O f g) ~> O (Companion hk f) (Companion hk g)

compFromCompose :: forall {j1 :: c} {j2 :: c} {k :: c} (f :: vk j2 k) (g :: vk j1 j2). Obj f -> Obj g -> O (Companion hk f) (Companion hk g) ~> Companion hk (O f g)

class Adjunction (Companion hk f) (Conjoint hk f) => ComConAdjunction (hk :: CAT c) (vk1 :: k) (f :: vk j k1)

class HasCompanions hk vk => Equipment (hk :: CAT c) (vk :: CAT c) | hk -> vk where

Minimal complete definition

mapConjoint, withObConjoint

Associated Types

type Conjoint (hk :: CAT c) (f :: vk j k) :: hk k j

Methods

mapConjoint :: forall {j :: c} {k :: c} (f :: vk j k) (g :: vk j k). (f ~> g) -> Conjoint hk g ~> Conjoint hk f

withObConjoint :: forall {j :: c} {k :: c} (f :: vk j k) r. Ob f => (Ob (Conjoint hk f) => r) -> r

conjToId :: forall (k :: c). Ob0 vk k => Conjoint hk (I :: vk k k) ~> (I :: hk k k)

conjFromId :: forall (k :: c). Ob0 vk k => (I :: hk k k) ~> Conjoint hk (I :: vk k k)

conjToCompose :: forall {k1 :: c} {j :: c} {k2 :: c} (f :: vk j k2) (g :: vk k1 j). Obj f -> Obj g -> Conjoint hk (O f g) ~> O (Conjoint hk g) (Conjoint hk f)

conjFromCompose :: forall {j1 :: c} {j2 :: c} {k :: c} (f :: vk j2 k) (g :: vk j1 j2). Obj f -> Obj g -> O (Conjoint hk g) (Conjoint hk f) ~> Conjoint hk (O f g)

comConUnit :: forall {j :: c} {k :: c} (f :: vk j k). Obj f -> (I :: hk j j) ~> O (Conjoint hk f) (Companion hk f)

default comConUnit :: forall {j :: c} {k :: c} (f :: vk j k). (Ob f => ComConAdjunction hk vk f) => Obj f -> (I :: hk j j) ~> O (Conjoint hk f) (Companion hk f)

comConCounit :: forall {j :: c} {k :: c} (f :: vk j k). Obj f -> O (Companion hk f) (Conjoint hk f) ~> (I :: hk k k)

default comConCounit :: forall {j :: c} {k :: c} (f :: vk j k). (Ob f => ComConAdjunction hk vk f) => Obj f -> O (Companion hk f) (Conjoint hk f) ~> (I :: hk k k)

type Cart (p :: hk h j) (f :: vk h i) (g :: vk j k) = O (O (Companion hk g) p) (Conjoint hk f)

P(f, g)

type SQ' (hk :: CAT c) (vk :: CAT c) (h :: c) (i :: c) (j :: c) (k :: c) = (hk k i, vk j k) +-> (hk j h, vk h i)

The kind of a square '(q, f) '(p, g).

h--f--i
|  v  |
p--@--q
|  v  |
j--g--k

type SQ (hk :: CAT c) (vk :: CAT c) = forall {h :: c} {i :: c} {j :: c} {k :: c}. SQ' hk vk h i j k

data Sq (pf :: (hk j h, vk h i)) (qg :: (hk k i, vk j k)) where

Constructors

Sq :: forall {c} {hk :: CAT c} {vk :: CAT c} {h :: c} {i :: c} {j :: c} {k :: c} (p :: hk j h) (q :: hk k i) (f :: vk h i) (g :: vk j k). (Ob0 vk h, Ob0 vk i, Ob0 vk j, Ob0 vk k, Ob p, Ob q, Ob f, Ob g) => (O (Companion hk f) p ~> O q (Companion hk g)) -> Sq '(p, f) '(q, g)

object :: forall {c} hk vk (k :: c). (HasCompanions hk vk, Ob0 vk k) => Sq '(I :: hk k k, I :: vk k k) '(I :: hk k k, I :: vk k k)

The empty square for an object.

k-----k
|     |
|     |
|     |
k-----k

hArr :: forall {c} {hk} {vk} {j :: c} {k :: c} (p :: hk j k) (q :: hk j k). (HasCompanions hk vk, Ob0 vk j, Ob0 vk k) => (p ~> q) -> Sq '(p, I :: vk k k) '(q, I :: vk j j)

Make a square from a horizontal arrow

k-----k
|     |
p--@--q
|     |
j-----j

hId :: forall {c} hk vk (j :: c) (k :: c) (p :: hk k j). (HasCompanions hk vk, Ob0 vk j, Ob0 vk k, Ob p) => Sq '(p, I :: vk j j) '(p, I :: vk k k)

compId :: forall {c} {hk} {vk} {j :: c} {k :: c} (f :: vk j k). (HasCompanions hk vk, Ob0 vk j, Ob0 vk k, Ob f) => Sq '(Companion hk f, I :: vk k k) '(Companion hk f, I :: vk j j)

conjId :: forall {c} {hk} {vk} {j :: c} {k :: c} (f :: vk j k). (Equipment hk vk, Ob0 vk j, Ob0 vk k, Ob f) => Sq '(Conjoint hk f, I :: vk j j) '(Conjoint hk f, I :: vk k k)

vArr :: forall {c} {hk} {vk} {j :: c} {k :: c} (f :: vk j k) (g :: vk j k). (HasCompanions hk vk, Ob0 vk j, Ob0 vk k) => (f ~> g) -> Sq '(I :: hk j j, f) '(I :: hk k k, g)

Make a square from a vertical arrow

j--f--k
|  v  |
|  @  |
|  v  |
j--g--k

vId :: forall {c} {hk} {vk} {j :: c} {k :: c} (f :: vk j k). (HasCompanions hk vk, Ob0 vk j, Ob0 vk k, Ob f) => Sq '(I :: hk j j, f) '(I :: hk k k, f)

(|||) :: forall {s} {j :: s} {k :: s} {i :: s} {hk} {vk} {h :: s} {l :: s} {m :: s} (p :: hk m l) (q :: hk j h) (r :: hk k i) (d :: vk l h) (e :: vk m j) (f :: vk h i) (g :: vk j k). HasCompanions hk vk => Sq '(p, d) '(q, e) -> Sq '(q, f) '(r, g) -> Sq '(p, O f d) '(r, O g e) infixl 6

Horizontal composition

l--d--h     h--f--i     l-f∘d-i
|  v  |     |  v  |     |  v  |
p--@--q ||| q--@--r  =  p--@--r
|  v  |     |  v  |     |  v  |
m--e--j     j--g--k     m-g∘e-k

(===) :: forall {s1} {k :: s1} {j1 :: s1} {hk} {vk} {h :: s1} {i :: s1} {j2 :: s1} {l :: s1} (p :: hk l j2) (q :: hk k j1) (r :: hk j2 h) (s2 :: hk j1 i) (e :: vk h i) (f :: vk j2 j1) (g :: vk l k). HasCompanions hk vk => Sq '(r, e) '(s2, f) -> Sq '(p, f) '(q, g) -> Sq '(O r p, e) '(O s2 q, g) infixl 5

Vertical composition

 h--e--i
 |  v  |
 r--@--s
 |  v  |
 j--f--k
   ===
 j--f--k
 |  v  |
 p--@--q
 |  v  |
 l--g--m

   v v

 h--e--i
 |  v  |
r∘p-@-s∘q
 |  v  |
 j--g--k

toRight :: forall {c} {hk} {vk} {j :: c} {k :: c} (f :: vk j k). (HasCompanions hk vk, Ob' f) => Sq '(I :: hk j j, f) '(Companion hk f, I :: vk j j)

fromLeft :: forall {c} {hk} {vk} {j :: c} {k :: c} (f :: vk j k). (HasCompanions hk vk, Ob' f) => Sq '(Companion hk f, I :: vk k k) '(I :: hk k k, f)

fromRight :: forall {c} {hk} {vk} {j :: c} {k :: c} (f :: vk j k). (Equipment hk vk, Ob0 vk j, Ob0 vk k, Ob f) => Sq '(I :: hk j j, I :: vk j j) '(Conjoint hk f, f)

toLeft :: forall {c} {hk} {vk} {j :: c} {k :: c} (f :: vk j k). (Equipment hk vk, Ob0 vk j, Ob0 vk k, Ob f) => Sq '(Conjoint hk f, f) '(I :: hk k k, I :: vk k k)

flipCompanion :: forall {s} {i :: s} {j :: s} {k :: s} {hk :: CAT s} {vk :: s -> s -> Type} (f :: vk j k) (p :: hk i j) (q :: hk i k). (Equipment hk vk, Ob p) => Obj f -> (O (Companion hk f) p ~> q) -> p ~> O (Conjoint hk f) q

flipCompanionInv :: forall {s} {i :: s} {j :: s} {k :: s} {hk :: CAT s} {vk :: s -> s -> Type} (f :: vk j k) (p :: hk i j) (q :: hk i k). (Equipment hk vk, Ob q) => Obj f -> (p ~> O (Conjoint hk f) q) -> O (Companion hk f) p ~> q

flipConjoint :: forall {c} {k1 :: c} {j :: c} {k2 :: c} {hk :: CAT c} {vk :: c -> c -> Type} (f :: vk j k2) (p :: hk j k1) (q :: hk k2 k1). (Equipment hk vk, Ob p) => Obj f -> (O p (Conjoint hk f) ~> q) -> p ~> O q (Companion hk f)

flipConjointInv :: forall {c} {k1 :: c} {j :: c} {k2 :: c} {hk :: CAT c} {vk :: c -> c -> Type} (f :: vk j k2) (p :: hk j k1) (q :: hk k2 k1). (Equipment hk vk, Ob q) => Obj f -> (p ~> O q (Companion hk f)) -> O p (Conjoint hk f) ~> q

data RetroSq (pf :: (hk k i, vk j k)) (qg :: (hk j h, vk h i)) where

The kind of a retro square '(q, g) '(p, f).

h--f--i
|  v  |
p----q
|  v  |
j--g--k

Constructors

RetroSq :: forall {c} {hk :: CAT c} {vk :: CAT c} {h :: c} {i :: c} {j :: c} {k :: c} (p :: hk j h) (q :: hk k i) (f :: vk h i) (g :: vk j k). (Ob0 vk h, Ob0 vk i, Ob0 vk j, Ob0 vk k, Ob p, Ob q, Ob f, Ob g) => (O q (Companion hk g) ~> O (Companion hk f) p) -> RetroSq '(q, g) '(p, f)