Proarrow.Category.Instance.Free

Documentation

type family All (cs :: [Kind -> Constraint]) k where ...

Equations

All ('[] :: [Kind -> Constraint]) k = ()

All (c ': cs) k = (c k, All cs k)

class (All cs k => c k) => FromAll (cs :: [Kind -> Constraint]) (c :: Kind -> Constraint) k

class (forall k. FromAll cs c k) => Elem (c :: Kind -> Constraint) (cs :: [Kind -> Constraint])

newtype FREE (cs :: [Kind -> Constraint]) (p :: CAT j)

Constructors

EMB j

data Free (a :: FREE cs p) (b :: FREE cs p) where

Constructors

Id :: forall {j} {cs :: [Kind -> Constraint]} {p :: CAT j} (a :: FREE cs p). Ob a => Free a a

Emb :: forall {j} (a1 :: j) (b1 :: j) (p :: CAT j) (cs :: [Kind -> Constraint]) (a :: FREE cs p). (Ob a1, Ob b1, Typeable a1, Typeable b1) => p a1 b1 -> Free a ('EMB a1 :: FREE cs p) -> Free a ('EMB b1 :: FREE cs p)

Str :: forall {j} {cs :: [Kind -> Constraint]} {p :: CAT j} (c :: Kind -> Constraint) (a1 :: FREE cs p) (b :: FREE cs p) (a :: FREE cs p). (HasStructure cs p c, Ob a1, Ob b) => Struct c a1 b -> Free a a1 -> Free a b

emb :: forall {j} (a :: j) (b :: j) (cs :: [Kind -> Constraint]) p. (Ob a, Ob b, Typeable a, Typeable b, Ok cs p) => p a b %1 -> Free ('EMB a :: FREE cs p) ('EMB b :: FREE cs p)

class (forall (x :: k) (y :: k1). Eq (p x y)) => Eq2 (p :: k -> k1 -> Type)

class (forall (x :: k) (y :: k1). Show (p x y)) => Show2 (p :: k -> k1 -> Type)

class (Typeable p, Typeable cs, Typeable j, All cs (FREE cs p)) => Ok (cs :: [Kind -> Constraint]) (p :: CAT j)

class (Ok c p, Eq2 p) => WithEq (a :: FREE c p)

class Show2 p => WithShow (a :: FREE c p)

showPostComp :: forall {j} {cs :: [Kind -> Constraint]} {p1 :: CAT j} p2 (a :: FREE cs p1) (b :: FREE cs p1). (Show p2, WithShow a) => Int -> p2 -> Free a b -> ShowS

class IsFreeOb (a :: FREE cs p) where

Associated Types

type Lower (f :: j +-> k) (a :: FREE cs p) :: k

Methods

withLowerOb :: forall {k} (f :: j +-> k) r. (Representable f, All cs k) => (Ob (Lower f a) => r) -> r

class ((Ok cs p, Eq2 p) => Eq2 str, Ok cs p => Typeable str, Show2 p => Show2 str) => CanEqShow (str :: CAT (FREE cs p))

class (Typeable c, CanEqShow (Struct c :: CAT (FREE cs p)), Elem c cs) => HasStructure (cs :: [Kind -> Constraint]) (p :: CAT j) (c :: Kind -> Constraint) where

Associated Types

data Struct (c :: Kind -> Constraint) :: CAT (FREE cs p)

Methods

foldStructure :: forall {k} (f :: j +-> k) (a :: FREE cs p) (b :: FREE cs p). (All cs k, Representable f) => (forall (x :: FREE cs p) (y :: FREE cs p). (x ~> y) -> Lower f x ~> Lower f y) -> Struct c a b -> Lower f a ~> Lower f b

fold :: forall {j} {k} {p} (cs :: [Kind -> Constraint]) (f :: j +-> k) (a :: FREE cs p) (b :: FREE cs p). (All cs k, Representable f) => (forall (x :: j) (y :: j). p x y -> (f % x) ~> (f % y)) -> (a ~> b) -> Lower f a ~> Lower f b

retract :: forall {j} {k} (cs :: [Kind -> Constraint]) (f :: j +-> k) (a :: FREE cs (InitialProfunctor :: j -> j -> Type)) (b :: FREE cs (InitialProfunctor :: j -> j -> Type)). (All cs k, Representable f) => (a ~> b) -> Lower f a ~> Lower f b