Proarrow.Core

Documentation

type PRO j k = j -> k -> Type

type CAT k = PRO k k

type BI k = (k, k) -> k

type OB k = k -> Constraint

class (Profunctor cat, (~>) ~ cat) => Category (cat :: CAT k) where

Associated Types

type Ob (a :: k) :: Constraint

type Ob a = ()

Methods

id :: Ob a => cat a a

(.) :: cat b c -> cat a b -> cat a c infixr 9

type family (~>) :: CAT k infixr 0

type CategoryOf k = Category ((~>) :: CAT k)

type (:~>) p q = forall a b. p a b ~> q a b infixr 0

class (CategoryOf j, CategoryOf k) => Profunctor (p :: PRO j k) where

Minimal complete definition

dimap

Methods

dimap :: (c ~> a) -> (b ~> d) -> p a b -> p c d

(\\) :: ((Ob a, Ob b) => r) -> p a b -> r infixl 1

default (\\) :: (Ob a, Ob b) => ((Ob a, Ob b) => r) -> p a b -> r

(//) :: Profunctor p => p a b -> ((Ob a, Ob b) => r) -> r infixr 0

lmap :: Profunctor p => (c ~> a) -> p a b -> p c b

rmap :: Profunctor p => (b ~> d) -> p a b -> p a d

dimapDefault :: CategoryOf k => ((c :: k) ~> a) -> ((b :: k) ~> d) -> (a ~> b) -> c ~> d

type family UN (w :: j -> k) (wa :: k) :: j

A helper type family to unwrap a wrapped kind. This is needed because the field selector functions of newtypes have to be lower case and therefore cannot be used at the type level.

type Is w a = a ~ w (UN w a)

Is w a checks that the kind a is a kind wrapped by w.