Proarrow.Core

Documentation

type PRO j k = j -> k -> Type

type (+->) j k = PRO k j infixr 0

type CAT k = PRO k k

type BI k = (k, k) -> k

type OB k = k -> Constraint

type Kind = Type

class Any (a :: k)

class Promonad ((~>) :: CAT k) => CategoryOf k

Associated Types

type (~>) :: CAT k infixr 0

type Ob (a :: k)

type Ob (a :: k) = Any a

class (Promonad cat, CategoryOf k, cat ~ ((~>) :: CAT k)) => Category (cat :: CAT k)

type (:~>) (p :: k -> k1 -> Type) (q :: k -> k1 -> Type) = forall (a :: k) (b :: k1). p a b -> q a b infixr 0

class (CategoryOf j, CategoryOf k) => Profunctor (p :: PRO j k) where

Minimal complete definition

dimap

Methods

dimap :: forall (c :: j) (a :: j) (b :: k) (d :: k). (c ~> a) -> (b ~> d) -> p a b -> p c d

(\\) :: forall (a :: j) (b :: k) r. ((Ob a, Ob b) => r) -> p a b -> r infixl 1

default (\\) :: forall (a :: j) (b :: k) r. (Ob a, Ob b) => ((Ob a, Ob b) => r) -> p a b -> r

(//) :: forall {k1} {k2} p (a :: k1) (b :: k2) r. Profunctor p => p a b -> ((Ob a, Ob b) => r) -> r infixr 0

lmap :: forall {j} {k} p (c :: j) (a :: j) (b :: k). Profunctor p => (c ~> a) -> p a b -> p c b

rmap :: forall {j} {k} p (b :: k) (d :: k) (a :: j). Profunctor p => (b ~> d) -> p a b -> p a d

dimapDefault :: forall {k} p (c :: k) (a :: k) (b :: k) (d :: k). Promonad p => p c a -> p b d -> p a b -> p c d

class Profunctor p => Promonad (p :: PRO k k) where

Methods

id :: forall (a :: k). Ob a => p a a

(.) :: forall (b :: k) (c :: k) (a :: k). p b c -> p a b -> p a c infixr 9

arr :: forall {k} p (a :: k) (b :: k). Promonad p => (a ~> b) -> p a b

type Obj (a :: k) = a ~> a

obj :: forall {k} (a :: k). (CategoryOf k, Ob a) => Obj a

src :: forall {k1} {k2} (a :: k2) (b :: k1) p. Profunctor p => p a b -> Obj a

tgt :: forall {k1} {k2} (a :: k2) (b :: k1) p. Profunctor p => p a b -> Obj b

type family UN (w :: j -> k) (wa :: k) :: j

A helper type family to unwrap a wrapped kind. This is needed because the field selector functions of newtypes have to be lower case and therefore cannot be used at the type level.

type Is (w :: j -> k) (a :: k) = a ~ w (UN w a)

Is w a checks that the kind a is a kind wrapped by w.