Proarrow.Functor

Documentation

type (.~>) (f :: k -> k1) (g :: k -> k1) = forall (a :: k). Ob a => f a ~> g a infixr 0

class (CategoryOf k1, CategoryOf k2, forall (a :: k1). Ob a => Ob' (f a)) => Functor (f :: k1 -> k2) where

Methods

map :: forall (a :: k1) (b :: k1). (a ~> b) -> f a ~> f b

newtype Prelude (f :: Type -> Type) a

Constructors

Prelude

Fields

unPrelude :: f a

newtype FromProfunctor (p :: k -> k1 -> Type) (a :: k) (b :: k1)

Constructors

FromProfunctor

Fields

unFromProfunctor :: p a b

type Presheaf k = () +-> k

Presheaves are functors but it makes more sense in proarrow to represent them as profunctors from the unit category.

type Copresheaf k = k +-> ()

Copresheaves are functors but it makes more sense in proarrow to represent them as profunctors into the unit category.